@@ -17,10 +17,9 @@
     "@types/react-dom": "18.0.9",
     "@xstate/react": "^3.2.2",
     "babylonjs": "^6.21.0",
-    "eslint-config-next": "13.0.5",
-    "next": "13.0.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-spring": "^9.5.5",
     "react-use-gesture": "^9.1.3",
     "storybook": "^6.5.13",
@@ -45,6 +44,7 @@
     "@typescript-eslint/parser": "^5.44.0",
     "babel-loader": "^8.3.0",
     "eslint": "^8.50.0",
+    "eslint-config-next": "^14.0.1",
     "eslint-config-standard-with-typescript": "^23.0.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jest": "^27.1.6",
