@@ -1,108 +1,861 @@
-export const Icons = {
-  Github: () => (
-    <svg
-      className="w-6 h-6 text-gray-800 dark:text-white"
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="currentColor"
-      viewBox="0 0 20 20"
-    >
-      <path
-        fillRule="evenodd"
-        d="M10 .333A9.911 9.911 0 0 0 6.866 19.65c.5.092.678-.215.678-.477 0-.237-.01-1.017-.014-1.845-2.757.6-3.338-1.169-3.338-1.169a2.627 2.627 0 0 0-1.1-1.451c-.9-.615.07-.6.07-.6a2.084 2.084 0 0 1 1.518 1.021 2.11 2.11 0 0 0 2.884.823c.044-.503.268-.973.63-1.325-2.2-.25-4.516-1.1-4.516-4.9A3.832 3.832 0 0 1 4.7 7.068a3.56 3.56 0 0 1 .095-2.623s.832-.266 2.726 1.016a9.409 9.409 0 0 1 4.962 0c1.89-1.282 2.717-1.016 2.717-1.016.366.83.402 1.768.1 2.623a3.827 3.827 0 0 1 1.02 2.659c0 3.807-2.319 4.644-4.525 4.889a2.366 2.366 0 0 1 .673 1.834c0 1.326-.012 2.394-.012 2.72 0 .263.18.572.681.475A9.911 9.911 0 0 0 10 .333Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  ),
-  Linkedin: () => (
-    <svg
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="currentColor"
-      viewBox="0 0 15 15"
-    >
-      <path
-        fillRule="evenodd"
-        d="M7.979 5v1.586a3.5 3.5 0 0 1 3.082-1.574C14.3 5.012 15 7.03 15 9.655V15h-3v-4.738c0-1.13-.229-2.584-1.995-2.584-1.713 0-2.005 1.23-2.005 2.5V15H5.009V5h2.97ZM3 2.487a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z"
-        clipRule="evenodd"
-      />
-      <path d="M3 5.012H0V15h3V5.012Z" />
-    </svg>
-  ),
-  Email: () => (
-    <svg
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 20 16"
-    >
-      <path
+type Icon = {
+  title: string;
+  ariaLabel: string;
+  icon: () => JSX.Element;
+};
+
+export const Icons: Record<string, Icon> = {
+  GitHub: {
+    title: "GitHub",
+    ariaLabel: "GitHub icon",
+    icon: () => (
+      <svg
+        className="w-6 h-6 text-gray-800 dark:text-white"
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        fill="currentColor"
+        viewBox="0 0 20 20"
+      >
+        <path
+          fillRule="evenodd"
+          d="M10 .333A9.911 9.911 0 0 0 6.866 19.65c.5.092.678-.215.678-.477 0-.237-.01-1.017-.014-1.845-2.757.6-3.338-1.169-3.338-1.169a2.627 2.627 0 0 0-1.1-1.451c-.9-.615.07-.6.07-.6a2.084 2.084 0 0 1 1.518 1.021 2.11 2.11 0 0 0 2.884.823c.044-.503.268-.973.63-1.325-2.2-.25-4.516-1.1-4.516-4.9A3.832 3.832 0 0 1 4.7 7.068a3.56 3.56 0 0 1 .095-2.623s.832-.266 2.726 1.016a9.409 9.409 0 0 1 4.962 0c1.89-1.282 2.717-1.016 2.717-1.016.366.83.402 1.768.1 2.623a3.827 3.827 0 0 1 1.02 2.659c0 3.807-2.319 4.644-4.525 4.889a2.366 2.366 0 0 1 .673 1.834c0 1.326-.012 2.394-.012 2.72 0 .263.18.572.681.475A9.911 9.911 0 0 0 10 .333Z"
+          clipRule="evenodd"
+        />
+      </svg>
+    ),
+  },
+  Linkedin: {
+    title: "Linkedin",
+    ariaLabel: "Linkedin icon",
+    icon: () => (
+      <svg
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        fill="currentColor"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fillRule="evenodd"
+          d="M7.979 5v1.586a3.5 3.5 0 0 1 3.082-1.574C14.3 5.012 15 7.03 15 9.655V15h-3v-4.738c0-1.13-.229-2.584-1.995-2.584-1.713 0-2.005 1.23-2.005 2.5V15H5.009V5h2.97ZM3 2.487a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z"
+          clipRule="evenodd"
+        />
+        <path d="M3 5.012H0V15h3V5.012Z" />
+      </svg>
+    ),
+  },
+  Email: {
+    title: "Email",
+    ariaLabel: "Email icon",
+    icon: () => (
+      <svg
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 20 16"
+      >
+        <path
+          stroke="currentColor"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          strokeWidth="2"
+          d="m19 2-8.4 7.05a1 1 0 0 1-1.2 0L1 2m18 0a1 1 0 0 0-1-1H2a1 1 0 0 0-1 1m18 0v12a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V2"
+        />
+      </svg>
+    ),
+  },
+  Download: {
+    title: "Download",
+    ariaLabel: "Download icon",
+    icon: () => (
+      <svg
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 20 19"
+      >
+        <path
+          stroke="currentColor"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          strokeWidth="2"
+          d="M15 15h.01M4 12H2a1 1 0 0 0-1 1v4a1 1 0 0 0 1 1h16a1 1 0 0 0 1-1v-4a1 1 0 0 0-1-1h-3M9.5 1v10.93m4-3.93-4 4-4-4"
+        />
+      </svg>
+    ),
+  },
+  PDF: {
+    title: "PDF",
+    ariaLabel: "PDF icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        xmlSpace="preserve"
+        id="_x32_"
+        width={200}
+        height={200}
+        fill="currentColor"
         stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="2"
-        d="m19 2-8.4 7.05a1 1 0 0 1-1.2 0L1 2m18 0a1 1 0 0 0-1-1H2a1 1 0 0 0-1 1m18 0v12a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V2"
-      />
-    </svg>
-  ),
-  Download: () => (
-    <svg
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 20 19"
-    >
-      <path
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="2"
-        d="M15 15h.01M4 12H2a1 1 0 0 0-1 1v4a1 1 0 0 0 1 1h16a1 1 0 0 0 1-1v-4a1 1 0 0 0-1-1h-3M9.5 1v10.93m4-3.93-4 4-4-4"
-      />
-    </svg>
-  ),
-  /** PDF  */
-  PDF: () => (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      xmlSpace="preserve"
-      id="_x32_"
-      width={200}
-      height={200}
-      fill="#FFF"
-      stroke="#FFF"
-      viewBox="0 0 512 512"
-    >
-      <g id="SVGRepo_iconCarrier">
-        <style>{".st0{fill:#fff}"}</style>
-        <path
-          d="M347.746 346.204c-8.398-.505-28.589.691-48.81 4.533-11.697-11.839-21.826-26.753-29.34-39.053 24.078-69.232 8.829-88.91-11.697-88.91-16.119 0-24.167 17.011-22.376 35.805.906 9.461 8.918 29.34 18.78 48.223-6.05 15.912-16.847 42.806-27.564 62.269-12.545 3.812-23.305 8.048-31.027 11.622-38.465 17.888-41.556 41.773-33.552 51.894 15.197 19.226 47.576 2.638 80.066-55.468 22.243-6.325 51.508-14.752 54.146-14.752.304 0 .721.097 1.204.253 16.215 14.298 35.366 30.67 51.128 32.825 22.808 3.136 35.791-13.406 34.891-23.692-.892-10.292-6.904-23.811-35.849-25.549zM203.761 408.88c-9.401 11.178-24.606 21.9-29.972 18.334-5.373-3.574-6.265-13.86 5.819-25.497 12.076-11.623 32.29-17.657 35.329-18.787 3.59-1.337 4.482 0 4.482 1.791s-6.265 12.968-15.658 24.159zm41.162-150.309c-.899-11.192 1.33-21.922 10.731-23.26 9.386-1.352 13.868 9.386 10.292 26.828-3.582 17.464-5.38 29.08-7.164 30.44-1.79 1.338-3.567-3.144-3.567-3.144-3.588-7.165-9.4-19.687-10.292-30.864zm3.582 105.126c4.912-8.064 17.442-40.702 17.442-40.702 2.683 4.926 23.699 29.956 23.699 29.956s-32.208 7.172-41.141 10.746zm97.494 14.298c-13.414-1.768-36.221-17.895-36.221-17.895-3.128-1.337 24.992-5.157 35.79-4.466 13.875.9 18.794 6.718 18.794 12.53 0 5.818-4.919 11.623-18.363 9.831z"
-          className="st0"
-        />
-        <path
-          d="m461.336 107.66-98.34-98.348L353.683 0H139.946C92.593 0 54.069 38.532 54.069 85.901v6.57H41.353v102.733h12.716v230.904c0 47.361 38.525 85.893 85.878 85.893h244.808c47.368 0 85.893-38.532 85.893-85.893V116.979l-9.312-9.319zm-76.582 372.533H139.946c-29.875 0-54.086-24.212-54.086-54.086V195.203h157.31V92.47H85.86V85.9c0-29.882 24.211-54.102 54.086-54.102H332.89v60.894c0 24.888 20.191 45.065 45.079 45.065h60.886v288.349c0 29.876-24.219 54.087-54.101 54.087zM88.09 166.086v-47.554c0-.839.683-1.524 1.524-1.524h15.108c2.49 0 4.786.409 6.837 1.212 2.029.795 3.812 1.91 5.299 3.322 1.501 1.419 2.653 3.144 3.433 5.121.78 1.939 1.182 4.058 1.182 6.294 0 2.282-.402 4.414-1.19 6.332a14.667 14.667 0 0 1-3.418 5.054c-1.479 1.427-3.27 2.549-5.321 3.329-2.036.78-4.332 1.174-6.822 1.174h-6.376v17.241c0 .84-.683 1.523-1.523 1.523h-7.208a1.526 1.526 0 0 1-1.525-1.524zm46.595 0v-47.554c0-.839.684-1.524 1.524-1.524h16.698c3.173 0 5.968.528 8.324 1.568 2.386 1.062 4.518 2.75 6.347 5.009a14.832 14.832 0 0 1 2.236 3.916 22.97 22.97 0 0 1 1.189 4.407c.253 1.531.401 3.181.453 4.957.045 1.694.067 3.515.067 5.447 0 1.924-.022 3.746-.067 5.44a37.452 37.452 0 0 1-.453 4.964 22.85 22.85 0 0 1-1.189 4.399 14.781 14.781 0 0 1-2.23 3.909c-1.842 2.282-3.976 3.969-6.354 5.016-2.334 1.04-5.135 1.568-8.324 1.568h-16.698a1.522 1.522 0 0 1-1.523-1.522zm79.584-28.105c.84 0 1.523.684 1.523 1.524v6.48c0 .84-.683 1.524-1.523 1.524h-18.244v18.579c0 .84-.684 1.523-1.524 1.523h-7.209c-.84 0-1.523-.683-1.523-1.523v-47.554c0-.839.683-1.524 1.523-1.524h27.653c.839 0 1.524.684 1.524 1.524v6.48c0 .84-.684 1.524-1.524 1.524h-18.92v11.444h18.244z"
-          className="st0"
-        />
-        <path
-          d="M109.418 137.706c1.212-1.092 1.798-2.645 1.798-4.749 0-2.096-.587-3.649-1.798-4.741-1.263-1.13-2.928-1.68-5.098-1.68h-5.975v12.848h5.975c2.169.001 3.835-.548 5.098-1.678zM156.139 157.481c1.13-.424 2.103-1.107 2.973-2.088.944-1.055 1.538-2.571 1.769-4.511.26-2.208.386-5.091.386-8.569 0-3.485-.126-6.369-.386-8.569-.231-1.946-.825-3.462-1.762-4.51-.869-.982-1.873-1.679-2.972-2.089-1.182-.453-2.534-.676-4.042-.676h-7.164v31.68h7.164c1.5.001 2.86-.222 4.034-.668z"
-          className="st0"
-        />
-      </g>
-    </svg>
-  ),
+        viewBox="0 0 512 512"
+      >
+        <g id="SVGRepo_iconCarrier">
+          <style>{".st0{fill:#fff}"}</style>
+          <path
+            d="M347.746 346.204c-8.398-.505-28.589.691-48.81 4.533-11.697-11.839-21.826-26.753-29.34-39.053 24.078-69.232 8.829-88.91-11.697-88.91-16.119 0-24.167 17.011-22.376 35.805.906 9.461 8.918 29.34 18.78 48.223-6.05 15.912-16.847 42.806-27.564 62.269-12.545 3.812-23.305 8.048-31.027 11.622-38.465 17.888-41.556 41.773-33.552 51.894 15.197 19.226 47.576 2.638 80.066-55.468 22.243-6.325 51.508-14.752 54.146-14.752.304 0 .721.097 1.204.253 16.215 14.298 35.366 30.67 51.128 32.825 22.808 3.136 35.791-13.406 34.891-23.692-.892-10.292-6.904-23.811-35.849-25.549zM203.761 408.88c-9.401 11.178-24.606 21.9-29.972 18.334-5.373-3.574-6.265-13.86 5.819-25.497 12.076-11.623 32.29-17.657 35.329-18.787 3.59-1.337 4.482 0 4.482 1.791s-6.265 12.968-15.658 24.159zm41.162-150.309c-.899-11.192 1.33-21.922 10.731-23.26 9.386-1.352 13.868 9.386 10.292 26.828-3.582 17.464-5.38 29.08-7.164 30.44-1.79 1.338-3.567-3.144-3.567-3.144-3.588-7.165-9.4-19.687-10.292-30.864zm3.582 105.126c4.912-8.064 17.442-40.702 17.442-40.702 2.683 4.926 23.699 29.956 23.699 29.956s-32.208 7.172-41.141 10.746zm97.494 14.298c-13.414-1.768-36.221-17.895-36.221-17.895-3.128-1.337 24.992-5.157 35.79-4.466 13.875.9 18.794 6.718 18.794 12.53 0 5.818-4.919 11.623-18.363 9.831z"
+            className="st0"
+          />
+          <path
+            d="m461.336 107.66-98.34-98.348L353.683 0H139.946C92.593 0 54.069 38.532 54.069 85.901v6.57H41.353v102.733h12.716v230.904c0 47.361 38.525 85.893 85.878 85.893h244.808c47.368 0 85.893-38.532 85.893-85.893V116.979l-9.312-9.319zm-76.582 372.533H139.946c-29.875 0-54.086-24.212-54.086-54.086V195.203h157.31V92.47H85.86V85.9c0-29.882 24.211-54.102 54.086-54.102H332.89v60.894c0 24.888 20.191 45.065 45.079 45.065h60.886v288.349c0 29.876-24.219 54.087-54.101 54.087zM88.09 166.086v-47.554c0-.839.683-1.524 1.524-1.524h15.108c2.49 0 4.786.409 6.837 1.212 2.029.795 3.812 1.91 5.299 3.322 1.501 1.419 2.653 3.144 3.433 5.121.78 1.939 1.182 4.058 1.182 6.294 0 2.282-.402 4.414-1.19 6.332a14.667 14.667 0 0 1-3.418 5.054c-1.479 1.427-3.27 2.549-5.321 3.329-2.036.78-4.332 1.174-6.822 1.174h-6.376v17.241c0 .84-.683 1.523-1.523 1.523h-7.208a1.526 1.526 0 0 1-1.525-1.524zm46.595 0v-47.554c0-.839.684-1.524 1.524-1.524h16.698c3.173 0 5.968.528 8.324 1.568 2.386 1.062 4.518 2.75 6.347 5.009a14.832 14.832 0 0 1 2.236 3.916 22.97 22.97 0 0 1 1.189 4.407c.253 1.531.401 3.181.453 4.957.045 1.694.067 3.515.067 5.447 0 1.924-.022 3.746-.067 5.44a37.452 37.452 0 0 1-.453 4.964 22.85 22.85 0 0 1-1.189 4.399 14.781 14.781 0 0 1-2.23 3.909c-1.842 2.282-3.976 3.969-6.354 5.016-2.334 1.04-5.135 1.568-8.324 1.568h-16.698a1.522 1.522 0 0 1-1.523-1.522zm79.584-28.105c.84 0 1.523.684 1.523 1.524v6.48c0 .84-.683 1.524-1.523 1.524h-18.244v18.579c0 .84-.684 1.523-1.524 1.523h-7.209c-.84 0-1.523-.683-1.523-1.523v-47.554c0-.839.683-1.524 1.523-1.524h27.653c.839 0 1.524.684 1.524 1.524v6.48c0 .84-.684 1.524-1.524 1.524h-18.92v11.444h18.244z"
+            className="st0"
+          />
+          <path
+            d="M109.418 137.706c1.212-1.092 1.798-2.645 1.798-4.749 0-2.096-.587-3.649-1.798-4.741-1.263-1.13-2.928-1.68-5.098-1.68h-5.975v12.848h5.975c2.169.001 3.835-.548 5.098-1.678zM156.139 157.481c1.13-.424 2.103-1.107 2.973-2.088.944-1.055 1.538-2.571 1.769-4.511.26-2.208.386-5.091.386-8.569 0-3.485-.126-6.369-.386-8.569-.231-1.946-.825-3.462-1.762-4.51-.869-.982-1.873-1.679-2.972-2.089-1.182-.453-2.534-.676-4.042-.676h-7.164v31.68h7.164c1.5.001 2.86-.222 4.034-.668z"
+            className="st0"
+          />
+        </g>
+      </svg>
+    ),
+  },
   /** SKILLS */
-  React: () => (
-    <svg
-      className="w-6 h-6 text-gray-800 dark:text-white"
-      aria-hidden="true"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="currentColor"
-      viewBox="0 0 20 18"
-    >
-      <path d="M19.718 9c0-1.429-1.339-2.681-3.467-3.5.029-.18.077-.37.1-.545.217-2.058-.273-3.543-1.379-4.182-1.235-.714-2.983-.186-4.751 1.239C8.45.589 6.7.061 5.468.773c-1.107.639-1.6 2.124-1.379 4.182.018.175.067.365.095.545C2.057 6.319.718 7.571.718 9c0 1.429 1.339 2.681 3.466 3.5-.028.18-.077.37-.095.545-.218 2.058.272 3.543 1.379 4.182.376.213.803.322 1.235.316a5.987 5.987 0 0 0 3.514-1.56 5.992 5.992 0 0 0 3.515 1.56 2.44 2.44 0 0 0 1.236-.316c1.106-.639 1.6-2.124 1.379-4.182-.019-.175-.067-.365-.1-.545 2.132-.819 3.471-2.071 3.471-3.5Zm-6.01-7.548a1.5 1.5 0 0 1 .76.187c.733.424 1.055 1.593.884 3.212-.012.106-.043.222-.058.33-.841-.243-1.7-.418-2.57-.523a16.165 16.165 0 0 0-1.747-1.972 4.9 4.9 0 0 1 2.731-1.234Zm-7.917 8.781c.172.34.335.68.529 1.017.194.337.395.656.6.969a14.09 14.09 0 0 1-1.607-.376 14.38 14.38 0 0 1 .478-1.61Zm-.479-4.076a14.085 14.085 0 0 1 1.607-.376c-.205.313-.405.634-.6.969-.195.335-.357.677-.529 1.017-.19-.527-.35-1.064-.478-1.61ZM6.3 9c.266-.598.563-1.182.888-1.75.33-.568.69-1.118 1.076-1.65.619-.061 1.27-.1 1.954-.1.684 0 1.333.035 1.952.1a19.63 19.63 0 0 1 1.079 1.654A19.3 19.3 0 0 1 14.136 9a18.869 18.869 0 0 1-1.953 3.403 19.218 19.218 0 0 1-3.931 0 20.163 20.163 0 0 1-1.066-1.653A19.33 19.33 0 0 1 6.3 9Zm7.816 2.25c.2-.337.358-.677.53-1.017.191.527.35 1.065.478 1.611a14.48 14.48 0 0 1-1.607.376c.202-.314.404-.635.597-.97h.002Zm.53-3.483c-.172-.34-.335-.68-.53-1.017a20.214 20.214 0 0 0-.6-.97c.542.095 1.078.22 1.606.376a14.113 14.113 0 0 1-.478 1.611h.002ZM10.217 3.34c.4.375.777.773 1.13 1.193-.37-.02-.746-.033-1.129-.033s-.76.013-1.131.033c.353-.42.73-.817 1.13-1.193Zm-4.249-1.7a1.5 1.5 0 0 1 .76-.187 4.9 4.9 0 0 1 2.729 1.233A16.25 16.25 0 0 0 7.71 4.658c-.87.105-1.728.28-2.569.524-.015-.109-.047-.225-.058-.331-.171-1.619.151-2.787.885-3.211ZM1.718 9c0-.9.974-1.83 2.645-2.506.218.857.504 1.695.856 2.506-.352.811-.638 1.65-.856 2.506C2.692 10.83 1.718 9.9 1.718 9Zm4.25 7.361c-.734-.423-1.056-1.593-.885-3.212.011-.106.043-.222.058-.331.84.243 1.697.418 2.564.524a16.37 16.37 0 0 0 1.757 1.982c-1.421 1.109-2.714 1.488-3.494 1.037Zm3.11-2.895c.374.021.753.034 1.14.034.387 0 .765-.013 1.139-.034a14.4 14.4 0 0 1-1.14 1.215 14.232 14.232 0 0 1-1.139-1.215Zm5.39 2.895c-.782.451-2.075.072-3.5-1.038a16.248 16.248 0 0 0 1.757-1.981 16.41 16.41 0 0 0 2.565-.523c.015.108.046.224.058.33.175 1.619-.148 2.789-.88 3.212Zm1.6-4.854A16.562 16.562 0 0 0 15.216 9c.352-.811.638-1.65.856-2.507C17.743 7.17 18.718 8.1 18.718 9c0 .9-.975 1.83-2.646 2.507h-.004Z" />
-      <path d="M10.215 10.773a1.792 1.792 0 1 0-1.786-1.8v.006a1.788 1.788 0 0 0 1.786 1.794Z" />
-    </svg>
-  ),
+  React: {
+    title: "React",
+    ariaLabel: "React icon",
+    icon: () => (
+      <svg
+        className="w-6 h-6 text-gray-800 dark:text-white"
+        aria-hidden="true"
+        xmlns="http://www.w3.org/2000/svg"
+        fill="currentColor"
+        viewBox="0 0 20 18"
+      >
+        <path d="M19.718 9c0-1.429-1.339-2.681-3.467-3.5.029-.18.077-.37.1-.545.217-2.058-.273-3.543-1.379-4.182-1.235-.714-2.983-.186-4.751 1.239C8.45.589 6.7.061 5.468.773c-1.107.639-1.6 2.124-1.379 4.182.018.175.067.365.095.545C2.057 6.319.718 7.571.718 9c0 1.429 1.339 2.681 3.466 3.5-.028.18-.077.37-.095.545-.218 2.058.272 3.543 1.379 4.182.376.213.803.322 1.235.316a5.987 5.987 0 0 0 3.514-1.56 5.992 5.992 0 0 0 3.515 1.56 2.44 2.44 0 0 0 1.236-.316c1.106-.639 1.6-2.124 1.379-4.182-.019-.175-.067-.365-.1-.545 2.132-.819 3.471-2.071 3.471-3.5Zm-6.01-7.548a1.5 1.5 0 0 1 .76.187c.733.424 1.055 1.593.884 3.212-.012.106-.043.222-.058.33-.841-.243-1.7-.418-2.57-.523a16.165 16.165 0 0 0-1.747-1.972 4.9 4.9 0 0 1 2.731-1.234Zm-7.917 8.781c.172.34.335.68.529 1.017.194.337.395.656.6.969a14.09 14.09 0 0 1-1.607-.376 14.38 14.38 0 0 1 .478-1.61Zm-.479-4.076a14.085 14.085 0 0 1 1.607-.376c-.205.313-.405.634-.6.969-.195.335-.357.677-.529 1.017-.19-.527-.35-1.064-.478-1.61ZM6.3 9c.266-.598.563-1.182.888-1.75.33-.568.69-1.118 1.076-1.65.619-.061 1.27-.1 1.954-.1.684 0 1.333.035 1.952.1a19.63 19.63 0 0 1 1.079 1.654A19.3 19.3 0 0 1 14.136 9a18.869 18.869 0 0 1-1.953 3.403 19.218 19.218 0 0 1-3.931 0 20.163 20.163 0 0 1-1.066-1.653A19.33 19.33 0 0 1 6.3 9Zm7.816 2.25c.2-.337.358-.677.53-1.017.191.527.35 1.065.478 1.611a14.48 14.48 0 0 1-1.607.376c.202-.314.404-.635.597-.97h.002Zm.53-3.483c-.172-.34-.335-.68-.53-1.017a20.214 20.214 0 0 0-.6-.97c.542.095 1.078.22 1.606.376a14.113 14.113 0 0 1-.478 1.611h.002ZM10.217 3.34c.4.375.777.773 1.13 1.193-.37-.02-.746-.033-1.129-.033s-.76.013-1.131.033c.353-.42.73-.817 1.13-1.193Zm-4.249-1.7a1.5 1.5 0 0 1 .76-.187 4.9 4.9 0 0 1 2.729 1.233A16.25 16.25 0 0 0 7.71 4.658c-.87.105-1.728.28-2.569.524-.015-.109-.047-.225-.058-.331-.171-1.619.151-2.787.885-3.211ZM1.718 9c0-.9.974-1.83 2.645-2.506.218.857.504 1.695.856 2.506-.352.811-.638 1.65-.856 2.506C2.692 10.83 1.718 9.9 1.718 9Zm4.25 7.361c-.734-.423-1.056-1.593-.885-3.212.011-.106.043-.222.058-.331.84.243 1.697.418 2.564.524a16.37 16.37 0 0 0 1.757 1.982c-1.421 1.109-2.714 1.488-3.494 1.037Zm3.11-2.895c.374.021.753.034 1.14.034.387 0 .765-.013 1.139-.034a14.4 14.4 0 0 1-1.14 1.215 14.232 14.232 0 0 1-1.139-1.215Zm5.39 2.895c-.782.451-2.075.072-3.5-1.038a16.248 16.248 0 0 0 1.757-1.981 16.41 16.41 0 0 0 2.565-.523c.015.108.046.224.058.33.175 1.619-.148 2.789-.88 3.212Zm1.6-4.854A16.562 16.562 0 0 0 15.216 9c.352-.811.638-1.65.856-2.507C17.743 7.17 18.718 8.1 18.718 9c0 .9-.975 1.83-2.646 2.507h-.004Z" />
+        <path d="M10.215 10.773a1.792 1.792 0 1 0-1.786-1.8v.006a1.788 1.788 0 0 0 1.786 1.794Z" />
+      </svg>
+    ),
+  },
+  Illustrator: {
+    title: "Illustrator",
+    ariaLabel: "Illustrator icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 24 24"
+      >
+        <path
+          fill="currentColor"
+          d="M15.932 10.143V15.5h-1.364v-5.357h1.364ZM14.492 8.647c0-.418.307-.737.758-.737.463 0 .748.319.76.737 0 .407-.298.726-.77.726-.451 0-.748-.319-.748-.726Z"
+        />
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="m12.634 15.5-.671-2.101H9.609L8.971 15.5H7.574L9.95 8.086h1.727l2.409 7.414h-1.452Zm-2.233-4.95-.572 1.826h1.903l-.583-1.826c-.116-.365-.217-.766-.31-1.133a60.256 60.256 0 0 0-.064-.253h-.022c-.038.147-.075.304-.114.465-.074.31-.151.632-.238.921Z"
+          clipRule="evenodd"
+        />
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="M17.258 2.833a47.721 47.721 0 0 0-10.516 0c-2.012.225-3.637 1.81-3.873 3.832a45.921 45.921 0 0 0 0 10.67c.236 2.022 1.86 3.607 3.873 3.832 3.466.387 7.05.387 10.516 0 2.012-.225 3.637-1.81 3.873-3.832a45.925 45.925 0 0 0 0-10.67c-.236-2.022-1.86-3.607-3.873-3.832Zm-10.35 1.49a46.22 46.22 0 0 1 10.184 0c1.33.15 2.395 1.199 2.55 2.517a44.421 44.421 0 0 1 0 10.32c-.155 1.318-1.22 2.368-2.55 2.516a46.216 46.216 0 0 1-10.184 0c-1.33-.148-2.395-1.198-2.55-2.516a44.421 44.421 0 0 1 0-10.32c.155-1.318 1.22-2.368 2.55-2.516Z"
+          clipRule="evenodd"
+        />
+      </svg>
+    ),
+  },
+  Photoshop: {
+    title: "Photoshop",
+    ariaLabel: "Photoshop icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 24 24"
+      >
+        <path
+          fill="currentColor"
+          d="M15.932 10.143V15.5h-1.364v-5.357h1.364ZM14.492 8.647c0-.418.307-.737.758-.737.463 0 .748.319.76.737 0 .407-.298.726-.77.726-.451 0-.748-.319-.748-.726Z"
+        />
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="m12.634 15.5-.671-2.101H9.609L8.971 15.5H7.574L9.95 8.086h1.727l2.409 7.414h-1.452Zm-2.233-4.95-.572 1.826h1.903l-.583-1.826c-.116-.365-.217-.766-.31-1.133a60.256 60.256 0 0 0-.064-.253h-.022c-.038.147-.075.304-.114.465-.074.31-.151.632-.238.921Z"
+          clipRule="evenodd"
+        />
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="M17.258 2.833a47.721 47.721 0 0 0-10.516 0c-2.012.225-3.637 1.81-3.873 3.832a45.921 45.921 0 0 0 0 10.67c.236 2.022 1.86 3.607 3.873 3.832 3.466.387 7.05.387 10.516 0 2.012-.225 3.637-1.81 3.873-3.832a45.925 45.925 0 0 0 0-10.67c-.236-2.022-1.86-3.607-3.873-3.832Zm-10.35 1.49a46.22 46.22 0 0 1 10.184 0c1.33.15 2.395 1.199 2.55 2.517a44.421 44.421 0 0 1 0 10.32c-.155 1.318-1.22 2.368-2.55 2.516a46.216 46.216 0 0 1-10.184 0c-1.33-.148-2.395-1.198-2.55-2.516a44.421 44.421 0 0 1 0-10.32c.155-1.318 1.22-2.368 2.55-2.516Z"
+          clipRule="evenodd"
+        />
+      </svg>
+    ),
+  },
+  Asperite: {
+    title: "Asperite",
+    ariaLabel: "Asperite icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <path d="M4.006 0v1.6h15.988V0zm15.988 1.6v1.6h1.6V1.6zm1.6 1.6v14.4h-1.6v1.6H4.006v-1.6h-1.6V3.2H.809v17.6h1.599v1.6h1.599V24h15.988v-1.6h1.6v-1.6h1.598V3.2zm-19.187 0h1.599V1.6h-1.6zm4.796 3.2v6.4h1.6V6.4zm7.995 0v6.4h1.599V6.4z" />
+      </svg>
+    ),
+  },
+  ASPnet: {
+    title: "ASP.net",
+    ariaLabel: "ASPnet icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlSpace="preserve"
+        width={800}
+        height={800}
+        viewBox="0 0 512 512"
+      >
+        <path d="M295.474 319.537c.826 4.135 1.24 11.37 1.24 21.707v80.997h-22.327v-80.129c0-9.097-.868-15.894-2.604-20.404-1.737-4.507-4.817-8.104-9.241-10.792-4.424-2.683-9.613-4.031-15.567-4.031-9.51 0-17.717 3.019-24.622 9.055-6.905 6.037-10.357 17.49-10.357 34.359v71.942h-22.327V290.512h20.094v18.73c9.675-14.472 23.65-21.707 41.925-21.707 7.938 0 15.236 1.426 21.893 4.279 6.657 2.853 11.64 6.595 14.947 11.226s5.622 10.13 6.946 16.497zM511.5 423.052c-6.284 1.323-11.908 1.984-16.869 1.984-8.104 0-14.389-1.281-18.854-3.845-4.465-2.563-7.607-5.934-9.427-10.109-1.819-4.172-2.729-12.962-2.729-26.358v-79.542h-48.475c.111.12.235.219.347.339 11.164 11.99 16.746 28.859 16.746 50.608 0 1.322-.042 3.308-.124 5.953h-98.239c.827 14.472 4.92 25.553 12.279 33.243 7.36 7.69 16.539 11.535 27.537 11.535 8.187 0 15.174-2.149 20.963-6.45 5.788-4.3 10.378-11.163 13.769-20.59l23.071 2.853c-3.639 13.479-10.378 23.939-20.219 31.382s-22.41 11.164-37.708 11.164c-19.268 0-34.545-5.934-45.832-17.8-11.288-11.862-16.932-28.509-16.932-49.926 0-22.162 5.706-39.362 17.117-51.601 11.412-12.238 26.214-18.357 44.406-18.357 2.199 0 4.354.095 6.454.281h84.839v-28.579h22.203v28.579h22.451v17.365h-22.451v80.782c0 6.367.393 10.461 1.178 12.28.786 1.819 2.067 3.267 3.846 4.341 1.777 1.075 4.32 1.613 7.628 1.613 2.481 0 5.747-.29 9.8-.869l3.225 19.724zm-102.828-79.327c-.993-11.081-3.805-19.392-8.435-24.932-7.112-8.601-16.332-12.9-27.661-12.9-10.254 0-18.875 3.432-25.862 10.295-6.987 6.864-10.854 16.043-11.598 27.537h73.556zM157.881 395.42c-4.035 0-7.459 1.435-10.279 4.354-2.82 2.898-4.217 6.405-4.217 10.548 0 4.035 1.397 7.529 4.217 10.473 2.82 2.944 6.244 4.425 10.279 4.425 4.139 0 7.662-1.48 10.56-4.425 2.906-2.943 4.35-6.438 4.35-10.473 0-4.044-1.443-7.55-4.35-10.489-2.899-2.958-6.421-4.413-10.56-4.413zm143.166-128.224c-11.577 5.086-26.172 7.629-43.785 7.629-30.928 0-52.345-5.954-64.253-17.862-.05-.05-.096-.103-.146-.152l5.604 14.913h-58.84l-9.101-30.018H66.737l-8.865 30.018H.5L68.846 89.883h61.292l45.653 121.494 50.461-3.161c1.158 8.683 3.514 15.3 7.07 19.847 5.789 7.361 14.058 11.039 24.808 11.039 8.021 0 14.203-1.879 18.544-5.644 4.342-3.763 6.513-8.125 6.513-13.086 0-4.714-2.067-8.932-6.202-12.652-4.135-3.722-13.728-7.235-28.777-10.544-24.642-5.538-42.214-12.899-52.716-22.078-10.585-9.18-15.877-20.88-15.877-35.104 0-9.344 2.708-18.172 8.125-26.482 5.416-8.311 13.562-14.841 24.436-19.598 10.874-4.755 25.779-7.133 44.716-7.133 23.237 0 40.954 4.323 53.151 12.962 12.197 8.642 19.453 22.39 21.769 41.243l-52.965 3.102c-1.405-8.187-4.362-14.141-8.869-17.862-4.506-3.721-10.729-5.581-18.667-5.581-6.533 0-11.453 1.385-14.761 4.155-3.308 2.771-4.961 6.14-4.961 10.108 0 2.896 1.364 5.502 4.093 7.814 2.646 2.4 8.931 4.633 18.854 6.698 24.56 5.293 42.152 10.649 52.778 16.063 10.626 5.416 18.358 12.137 23.195 20.156a49.519 49.519 0 0 1 5.521 13.079V89.883h93.401c20.342 0 35.578 4.837 45.708 14.512 10.13 9.676 15.195 23.443 15.195 41.306 0 18.357-5.521 32.705-16.56 43.041-11.04 10.339-27.889 15.505-50.546 15.505h-30.762v67.478h-56.438v-44.868c-1.574 6.286-4.216 12.271-7.939 17.952-6.451 9.84-15.465 17.302-27.042 22.387zm91.417-99.789h13.769c10.832 0 18.44-1.88 22.823-5.645 4.383-3.762 6.573-8.577 6.573-14.45 0-5.706-1.901-10.543-5.705-14.513-3.804-3.969-10.957-5.954-21.459-5.954h-16.001v40.562zm-273.593 34.979-20.07-65.368-19.863 65.368h39.933z" />
+      </svg>
+    ),
+  },
+  AWS: {
+    title: "AWS",
+    ariaLabel: "AWS icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 16 16"
+      >
+        <g fill="currentColor">
+          <path d="M4.51 7.687c0 .197.02.357.058.475.042.117.096.245.17.384a.233.233 0 0 1 .037.123c0 .053-.032.107-.1.16l-.336.224a.255.255 0 0 1-.138.048c-.054 0-.107-.026-.16-.074a1.652 1.652 0 0 1-.192-.251 4.137 4.137 0 0 1-.164-.315c-.416.491-.937.737-1.565.737-.447 0-.804-.129-1.064-.385-.261-.256-.394-.598-.394-1.025 0-.454.16-.822.484-1.1.325-.278.756-.416 1.304-.416.18 0 .367.016.564.042.197.027.4.07.612.118v-.39c0-.406-.085-.689-.25-.854-.17-.166-.458-.246-.868-.246-.186 0-.377.022-.574.07a4.23 4.23 0 0 0-.575.181 1.525 1.525 0 0 1-.186.07.326.326 0 0 1-.085.016c-.075 0-.112-.054-.112-.166v-.262c0-.085.01-.15.037-.186a.399.399 0 0 1 .15-.113c.185-.096.409-.176.67-.24.26-.07.537-.101.83-.101.633 0 1.096.144 1.394.432.293.288.442.726.442 1.314v1.73h.01zm-2.161.811c.175 0 .356-.032.548-.096.192-.064.362-.182.505-.342a.848.848 0 0 0 .181-.341c.032-.129.054-.283.054-.465V7.03a4.43 4.43 0 0 0-.49-.09 3.996 3.996 0 0 0-.5-.033c-.357 0-.617.07-.793.214-.176.144-.26.347-.26.614 0 .25.063.437.196.566.128.133.314.197.559.197zm4.273.577c-.096 0-.16-.016-.202-.054-.043-.032-.08-.106-.112-.208l-1.25-4.127a.938.938 0 0 1-.048-.214c0-.085.042-.133.127-.133h.522c.1 0 .17.016.207.053.043.032.075.107.107.208l.894 3.535.83-3.535c.026-.106.058-.176.101-.208a.365.365 0 0 1 .213-.053h.426c.1 0 .17.016.212.053.043.032.08.107.102.208l.84 3.578.92-3.578a.459.459 0 0 1 .107-.208.347.347 0 0 1 .208-.053h.495c.085 0 .133.043.133.133 0 .027-.006.054-.01.086a.768.768 0 0 1-.038.133l-1.283 4.127c-.031.107-.069.177-.111.209a.34.34 0 0 1-.203.053h-.457c-.101 0-.17-.016-.213-.053-.043-.038-.08-.107-.101-.214L8.213 5.37l-.82 3.439c-.026.107-.058.176-.1.213-.043.038-.118.054-.213.054h-.458zm6.838.144a3.51 3.51 0 0 1-.82-.096c-.266-.064-.473-.134-.612-.214-.085-.048-.143-.101-.165-.15a.38.38 0 0 1-.031-.149v-.272c0-.112.042-.166.122-.166a.3.3 0 0 1 .096.016c.032.011.08.032.133.054.18.08.378.144.585.187.213.042.42.064.633.064.336 0 .596-.059.777-.176a.575.575 0 0 0 .277-.508.52.52 0 0 0-.144-.373c-.095-.102-.276-.193-.537-.278l-.772-.24c-.388-.123-.676-.305-.851-.545a1.275 1.275 0 0 1-.266-.774c0-.224.048-.422.143-.593.096-.17.224-.32.384-.438a1.7 1.7 0 0 1 .553-.277c.213-.064.436-.091.67-.091.118 0 .24.005.357.021a3.4 3.4 0 0 1 .649.145c.096.032.17.064.224.096a.461.461 0 0 1 .16.133.289.289 0 0 1 .047.176v.251c0 .112-.042.171-.122.171a.552.552 0 0 1-.202-.064 2.428 2.428 0 0 0-1.022-.208c-.303 0-.543.048-.708.15-.165.1-.25.256-.25.475 0 .149.053.277.16.379.106.101.303.202.585.293l.756.24c.383.123.66.294.825.513.165.219.244.47.244.748 0 .23-.047.437-.138.619a1.435 1.435 0 0 1-.388.47c-.165.133-.362.23-.591.299a2.52 2.52 0 0 1-.761.112z" />
+          <path
+            fillRule="evenodd"
+            d="M14.465 11.813c-1.75 1.297-4.294 1.986-6.481 1.986-3.065 0-5.827-1.137-7.913-3.027-.165-.15-.016-.353.18-.235 2.257 1.313 5.04 2.109 7.92 2.109 1.941 0 4.075-.406 6.039-1.239.293-.133.543.192.255.406z"
+            clipRule="evenodd"
+          />
+          <path
+            fillRule="evenodd"
+            d="M15.194 10.98c-.223-.287-1.479-.138-2.048-.069-.17.022-.197-.128-.043-.24 1-.705 2.645-.502 2.836-.267.192.24-.053 1.89-.99 2.68-.143.123-.281.06-.217-.1.212-.53.686-1.72.462-2.003z"
+            clipRule="evenodd"
+          />
+        </g>
+      </svg>
+    ),
+  },
+  Azure: {
+    title: "Azure",
+    ariaLabel: "Azure icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          strokeLinejoin="round"
+          d="m3.5 4.5-3 7h3l4-11-4 4ZM14.5 13.5l-5-11-2 5 3 4-6 2h10Z"
+        />
+      </svg>
+    ),
+  },
+  Bitbucket: {
+    title: "Bitbucket",
+    ariaLabel: "Bitbucket icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="M.5.5V0a.5.5 0 0 0-.495.57L.5.5Zm14 0 .495.07A.5.5 0 0 0 14.5 0v.5Zm-2 14v.5a.5.5 0 0 0 .495-.43l-.495-.07Zm-10 0-.495.07A.5.5 0 0 0 2.5 15v-.5ZM5 4.5V4a.5.5 0 0 0-.498.542L5 4.5Zm4.5 6v.5a.5.5 0 0 0 .498-.459L9.5 10.5Zm-4 0-.498.041A.5.5 0 0 0 5.5 11v-.5ZM.5 1h14V0H.5v1ZM14.005.43l-2 14 .99.14 2-14-.99-.14ZM12.5 14h-10v1h10v-1Zm-9.505.43-2-14-.99.14 2 14 .99-.14ZM5 5h5V4H5v1Zm4.502-.542-.5 6 .996.083.5-6-.996-.083ZM9.5 10h-4v1h4v-1Zm-3.502.459-.5-6-.996.083.5 6 .996-.083ZM10 5h4V4h-4v1Z"
+        />
+      </svg>
+    ),
+  },
+  Blender: {
+    title: "Blender",
+    ariaLabel: "Blender icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlSpace="preserve"
+        width={800}
+        height={800}
+        viewBox="0 0 512 512"
+      >
+        <path d="M510.003 279.642c-2.998-21.097-10.305-41.104-21.724-59.459-9.96-16.019-22.739-30.266-37.992-42.375l.041-.038L290.133 54.731a4.569 4.569 0 0 0-.361-.287c-5.327-4.08-12.538-6.325-20.298-6.325-7.769 0-15.262 2.25-21.088 6.337-6.263 4.375-9.843 10.18-10.093 16.36-.229 5.764 2.521 11.311 7.764 15.636 10.31 8.134 20.598 16.447 30.898 24.768 9.997 8.08 20.298 16.401 30.549 24.502l-196.213-.133c-22.439 0-37.718 10.537-40.861 28.178-1.381 7.727 1.056 16.223 6.504 22.73 5.78 6.899 14.172 10.704 23.629 10.704l14.958.009c20.664 0 41.419-.05 62.146-.1l19.766-.046L19.354 328.812l-.707.516C8.7 336.954 2.188 347.642.783 358.653c-1.065 8.342.881 15.965 5.63 22.053 5.66 7.257 14.497 11.249 24.885 11.249 10.205 0 20.618-3.867 29.334-10.908l96.166-78.7c-.411 3.843-.91 9.481-.853 13.573.108 6.479 2.188 19.479 5.481 30.033 6.804 21.691 18.265 41.536 34.063 58.964 16.438 18.131 36.458 32.508 59.5 42.721 24.361 10.774 50.547 16.243 77.836 16.243h.253c27.376-.066 53.646-5.622 78.085-16.518 23.08-10.334 43.091-24.769 59.467-42.899 15.778-17.516 27.222-37.394 34.013-59.067a151.126 151.126 0 0 0 6.417-33.003c.839-10.829.478-21.849-1.057-32.752zM334.821 383.601c-60.142 0-108.912-43.627-108.912-97.447 0-53.815 48.771-97.442 108.912-97.442 60.141 0 108.907 43.627 108.907 97.442.001 53.82-48.766 97.447-108.907 97.447zm62.806-106.01c.886 16.064-5.53 30.978-16.796 42.019-11.461 11.248-27.816 18.313-46.102 18.313-18.281 0-34.637-7.065-46.102-18.313-11.262-11.041-17.665-25.954-16.784-42.006.865-15.603 8.476-29.376 19.94-39.128 11.273-9.589 26.411-15.44 42.945-15.44 16.538 0 31.671 5.852 42.945 15.44 11.467 9.752 19.081 23.516 19.954 39.115z" />
+      </svg>
+    ),
+  },
+  Bootstrap: {
+    title: "Bootstrap",
+    ariaLabel: "Bootstrap icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 24 24"
+      >
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="M4.985 2c-1.37 0-2.383 1.199-2.337 2.498.043 1.25-.013 2.867-.42 4.186-.41 1.322-1.1 2.16-2.228 2.268v1.215c1.128.107 1.819.945 2.227 2.268.408 1.319.464 2.936.42 4.185-.045 1.3.968 2.499 2.338 2.499h14.032c1.37 0 2.383-1.199 2.337-2.499-.043-1.249.013-2.866.42-4.185.409-1.323 1.098-2.16 2.226-2.268v-1.215c-1.128-.108-1.817-.946-2.226-2.268-.407-1.32-.463-2.937-.42-4.186C21.4 3.198 20.386 2 19.017 2H4.985ZM16.27 13.769c0 1.79-1.335 2.875-3.55 2.875H8.949a.407.407 0 0 1-.407-.407V6.881a.407.407 0 0 1 .407-.406h3.75c1.847 0 3.06 1 3.06 2.537 0 1.078-.816 2.043-1.855 2.213v.056c1.415.155 2.367 1.135 2.367 2.488ZM12.31 7.764h-2.15v3.038h1.811c1.4 0 2.172-.564 2.172-1.572 0-.944-.664-1.466-1.833-1.466Zm-2.15 4.243v3.347h2.23c1.457 0 2.23-.585 2.23-1.684 0-1.1-.794-1.663-2.324-1.663h-2.136Z"
+          clipRule="evenodd"
+        />
+      </svg>
+    ),
+  },
+  Csharp: {
+    title: "C#",
+    ariaLabel: "Csharp icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M9.5 5v5m2-5v5M8 6.5h5m-5 2h5M8 10l-.402.201A2.831 2.831 0 0 1 3.5 7.668v-.336a2.832 2.832 0 0 1 4.098-2.533L8 5m-6.5 5.5v-6l6-3.5 6 3.5v6l-6 3.5-6-3.5Z"
+        />
+      </svg>
+    ),
+  },
+  ChakraUI: {
+    title: "ChakraUI",
+    ariaLabel: "ChakraUI icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <path d="M12 0C5.352 0 0 5.352 0 12s5.352 12 12 12 12-5.352 12-12S18.648 0 12 0zm2.8 4.333c.13-.004.248.136.171.278l-3.044 5.58a.187.187 0 0 0 .164.276h5.26c.17 0 .252.207.128.323l-9.22 8.605c-.165.154-.41-.063-.278-.246l4.364-6.021a.187.187 0 0 0-.151-.296H6.627a.187.187 0 0 1-.131-.32l8.18-8.123a.182.182 0 0 1 .125-.056z" />
+      </svg>
+    ),
+  },
+  CSS: {
+    title: "CSS",
+    ariaLabel: "CSS icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="M.5.5V0a.5.5 0 0 0-.498.542L.5.5Zm14 0 .498.042A.5.5 0 0 0 14.5 0v.5Zm-1 12 .158.474a.5.5 0 0 0 .34-.433L13.5 12.5Zm-6 2-.158.474a.499.499 0 0 0 .316 0L7.5 14.5Zm-6-2-.498.041a.5.5 0 0 0 .34.433L1.5 12.5Zm9-9h.5V3h-.5v.5Zm0 6 .158.474L11 9.86V9.5h-.5Zm-3 1-.158.474.158.053.158-.053L7.5 10.5Zm-3-1H4v.36l.342.114L4.5 9.5ZM.5 1h14V0H.5v1ZM14.002.458l-1 12 .996.083 1-12-.996-.083Zm-.66 11.568-6 2 .316.948 6-2-.316-.948Zm-5.684 2-6-2-.316.948 6 2 .316-.948Zm-5.66-1.567-1-12-.996.083 1 12 .996-.083ZM10.5 3H4v1h6.5V3ZM6 7h4.5V6H6v1Zm4-.5v3h1v-3h-1Zm.342 2.526-3 1 .316.948 3-1-.316-.948Zm-2.684 1-3-1-.316.948 3 1 .316-.948ZM5 9.5V8H4v1.5h1Zm5-6v3h1v-3h-1Z"
+        />
+      </svg>
+    ),
+  },
+  SQL: {
+    title: "SQL",
+    ariaLabel: "SQL icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          strokeLinecap="square"
+          d="M14.5 2.499c0 1.103-3.134 1.998-7 1.998S.5 3.602.5 2.5m14 0C14.5 1.395 11.366.5 7.5.5s-7 .895-7 1.999m14 0v9.993c0 1.103-3.134 1.999-7 1.999s-7-.895-7-1.999V2.5m14 4.996c0 1.104-3.134 2-7 2s-7-.896-7-2"
+        />
+      </svg>
+    ),
+  },
+  Docker: {
+    title: "Docker",
+    ariaLabel: "Docker icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="M.5 5.5V5H0v.5h.5Zm2-2V3H2v.5h.5Zm4-2V1H6v.5h.5Zm2 0H9V1h-.5v.5Zm4 6H12V8h.5v-.5ZM1 7.5v-2H0v2h1Zm2 0v-4H2v4h1ZM2.5 4h6V3h-6v1ZM8 3.5v4h1v-4H8Zm-3 4v-4H4v4h1Zm2 0v-6H6v6h1ZM6.5 2h2V1h-2v1ZM8 1.5v2h1v-2H8Zm5.736 8.5H15V9h-1.264v1ZM10 5v.5h1V5h-1Zm2 1.5v1h1v-1h-1Zm.5 1.5h1V7h-1v1Zm1.5.5v1h1v-1h-1Zm-.5-.5a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 13.5 7v1Zm-2-2a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 11.5 5v1ZM3 10h1V9H3v1Zm5.5-3h-8v1h8V7ZM0 7.5v1h1v-1H0ZM5.5 14h.528v-1H5.5v1Zm.528 0a7.736 7.736 0 0 0 6.23-3.15l-.805-.593A6.737 6.737 0 0 1 6.028 13v1ZM0 8.5A5.5 5.5 0 0 0 5.5 14v-1A4.5 4.5 0 0 1 1 8.5H0ZM.5 6h11V5H.5v1Zm9.5-.5A1.5 1.5 0 0 1 8.5 7v1A2.5 2.5 0 0 0 11 5.5h-1ZM13.736 9c-.96 0-1.769.558-2.283 1.257l.806.593c.383-.522.922-.85 1.477-.85V9Z"
+        />
+      </svg>
+    ),
+  },
+  Figma: {
+    title: "Figma",
+    ariaLabel: "Figma icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M7.5 1v8.5m0 0v2a2 2 0 1 1-2-2m2 0h-2m0 0a2 2 0 1 1 0-4m0 0h2m-2 0h4m-4 0a2 2 0 1 1 0-4h4a2 2 0 1 1 0 4m0 0a2 2 0 1 1 0 4 2 2 0 0 1 0-4Z"
+        />
+      </svg>
+    ),
+  },
+  GitHub2: {
+    title: "GitHub",
+    ariaLabel: "Github icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="M5.65 12.477a.5.5 0 1 0-.3-.954l.3.954Zm-3.648-2.96-.484-.128-.254.968.484.127.254-.968ZM9 14.5v.5h1v-.5H9Zm.063-4.813-.054-.497a.5.5 0 0 0-.299.852l.352-.354ZM12.5 5.913h.5V5.91l-.5.002Zm-.833-2.007-.466-.18a.5.5 0 0 0 .112.533l.354-.353Zm-.05-2.017.456-.204a.5.5 0 0 0-.319-.276l-.137.48Zm-2.173.792-.126.484a.5.5 0 0 0 .398-.064l-.272-.42Zm-3.888 0-.272.42a.5.5 0 0 0 .398.064l-.126-.484ZM3.383 1.89l-.137-.48a.5.5 0 0 0-.32.276l.457.204Zm-.05 2.017.354.353a.5.5 0 0 0 .112-.534l-.466.181ZM2.5 5.93H3v-.002l-.5.002Zm3.438 3.758.352.355a.5.5 0 0 0-.293-.851l-.06.496ZM5.5 11H6l-.001-.037L5.5 11ZM5 14.5v.5h1v-.5H5Zm.35-2.977c-.603.19-.986.169-1.24.085-.251-.083-.444-.25-.629-.49a4.8 4.8 0 0 1-.27-.402c-.085-.139-.182-.302-.28-.447-.191-.281-.473-.633-.929-.753l-.254.968c.08.02.184.095.355.346.082.122.16.252.258.412.094.152.202.32.327.484.253.33.598.663 1.11.832.51.168 1.116.15 1.852-.081l-.3-.954Zm4.65-.585c0-.318-.014-.608-.104-.878-.096-.288-.262-.51-.481-.727l-.705.71c.155.153.208.245.237.333.035.105.053.254.053.562h1Zm-.884-.753c.903-.097 1.888-.325 2.647-.982.78-.675 1.237-1.729 1.237-3.29h-1c0 1.359-.39 2.1-.892 2.534-.524.454-1.258.653-2.099.743l.107.995ZM13 5.91a3.354 3.354 0 0 0-.98-2.358l-.707.706c.438.44.685 1.034.687 1.655l1-.003Zm-.867-1.824c.15-.384.22-.794.21-1.207l-1 .025a2.12 2.12 0 0 1-.142.82l.932.362Zm.21-1.207a3.119 3.119 0 0 0-.27-1.195l-.913.408c.115.256.177.532.184.812l1-.025Zm-.726-.99c.137-.481.137-.482.136-.482h-.003l-.004-.002a.462.462 0 0 0-.03-.007 1.261 1.261 0 0 0-.212-.024 2.172 2.172 0 0 0-.51.054c-.425.091-1.024.317-1.82.832l.542.84c.719-.464 1.206-.634 1.488-.694.14-.03.23-.033.273-.032.022 0 .033.002.033.002l-.008-.001a.278.278 0 0 1-.01-.002l-.006-.002h-.003l-.002-.001c-.001 0-.002 0 .136-.482Zm-2.047.307a8.209 8.209 0 0 0-4.14 0l.252.968a7.209 7.209 0 0 1 3.636 0l.252-.968Zm-3.743.064C5.03 1.746 4.43 1.52 4.005 1.43a2.17 2.17 0 0 0-.51-.053 1.259 1.259 0 0 0-.241.03l-.004.002h-.003l.136.481.137.481h-.001l-.002.001-.003.001a.327.327 0 0 1-.016.004l-.008.001h.008a1.19 1.19 0 0 1 .298.03c.282.06.769.23 1.488.694l.543-.84Zm-2.9-.576a3.12 3.12 0 0 0-.27 1.195l1 .025c.006-.28.068-.556.183-.812l-.913-.408Zm-.27 1.195c-.01.413.06.823.21 1.207l.932-.362a2.12 2.12 0 0 1-.143-.82l-1-.025Zm.322.673a3.354 3.354 0 0 0-.726 1.091l.924.38c.118-.285.292-.545.51-.765l-.708-.706Zm-.726 1.091A3.354 3.354 0 0 0 2 5.93l1-.003c0-.31.06-.616.177-.902l-.924-.38ZM2 5.93c0 1.553.458 2.597 1.239 3.268.757.65 1.74.88 2.64.987l.118-.993C5.15 9.09 4.416 8.89 3.89 8.438 3.388 8.007 3 7.276 3 5.928H2Zm3.585 3.404c-.5.498-.629 1.09-.584 1.704L6 10.963c-.03-.408.052-.683.291-.921l-.705-.709ZM5 11v3.5h1V11H5Zm5 3.5V13H9v1.5h1Zm0-1.5v-2.063H9V13h1Z"
+        />
+      </svg>
+    ),
+  },
+  Gitlab: {
+    title: "Gitlab",
+    ariaLabel: "Gitlab icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          strokeLinejoin="round"
+          d="m.5 8.5 7 6 7-6-2-8-2 6h-6l-2-6-2 8Z"
+        />
+      </svg>
+    ),
+  },
+  GoLang: {
+    title: "GoLang",
+    ariaLabel: "GoLang icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <path d="M1.811 10.231c-.047 0-.058-.023-.035-.059l.246-.315c.023-.035.081-.058.128-.058h4.172c.046 0 .058.035.035.07l-.199.303c-.023.036-.082.07-.117.07zM.047 11.306c-.047 0-.059-.023-.035-.058l.245-.316c.023-.035.082-.058.129-.058h5.328c.047 0 .07.035.058.07l-.093.28c-.012.047-.058.07-.105.07zm2.828 1.075c-.047 0-.059-.035-.035-.07l.163-.292c.023-.035.07-.07.117-.07h2.337c.047 0 .07.035.07.082l-.023.28c0 .047-.047.082-.082.082zm12.129-2.36c-.736.187-1.239.327-1.963.514-.176.046-.187.058-.34-.117-.174-.199-.303-.327-.548-.444-.737-.362-1.45-.257-2.115.175-.795.514-1.204 1.274-1.192 2.22.011.935.654 1.706 1.577 1.835.795.105 1.46-.175 1.987-.77.105-.13.198-.27.315-.434H10.47c-.245 0-.304-.152-.222-.35.152-.362.432-.97.596-1.274a.315.315 0 0 1 .292-.187h4.253c-.023.316-.023.631-.07.947a4.983 4.983 0 0 1-.958 2.29c-.841 1.11-1.94 1.8-3.33 1.986-1.145.152-2.209-.07-3.143-.77-.865-.655-1.356-1.52-1.484-2.595-.152-1.274.222-2.419.993-3.424.83-1.086 1.928-1.776 3.272-2.02 1.098-.2 2.15-.07 3.096.571.62.41 1.063.97 1.356 1.648.07.105.023.164-.117.2m3.868 6.461c-1.064-.024-2.034-.328-2.852-1.029a3.665 3.665 0 0 1-1.262-2.255c-.21-1.32.152-2.489.947-3.529.853-1.122 1.881-1.706 3.272-1.95 1.192-.21 2.314-.095 3.33.595.923.63 1.496 1.484 1.648 2.605.198 1.578-.257 2.863-1.344 3.962-.771.783-1.718 1.273-2.805 1.495-.315.06-.63.07-.934.106zm2.78-4.72c-.011-.153-.011-.27-.034-.387-.21-1.157-1.274-1.81-2.384-1.554-1.087.245-1.788.935-2.045 2.033-.21.912.234 1.835 1.075 2.21.643.28 1.285.244 1.905-.07.923-.48 1.425-1.228 1.484-2.233z" />
+      </svg>
+    ),
+  },
+  GraphQL: {
+    title: "GraphQL",
+    ariaLabel: "GraphQL icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="m2.61 4.432 4.142-2.417-.504-.864-4.143 2.417.504.864ZM2 9.5v-4H1v4h1Zm6.248-7.485 4.143 2.417.504-.864-4.143-2.417-.504.864ZM13 5.5v4h1v-4h-1Zm-.252 4.86-4.5 2.625.504.864 4.5-2.625-.504-.864Zm-5.996 2.625-4.143-2.417-.504.864 4.143 2.417.504-.864ZM6.584 1.973l-5 7.5.832.554 5-7.5-.832-.554Zm6.832 7.5-5-7.5-.832.554 5 7.5.832-.554ZM2.5 11h10v-1h-10v1Zm5-9a.5.5 0 0 1-.5-.5H6A1.5 1.5 0 0 0 7.5 3V2Zm.5-.5a.5.5 0 0 1-.5.5v1A1.5 1.5 0 0 0 9 1.5H8ZM7.5 1a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 7.5 0v1Zm0-1A1.5 1.5 0 0 0 6 1.5h1a.5.5 0 0 1 .5-.5V0Zm6 5a.5.5 0 0 1-.5-.5h-1A1.5 1.5 0 0 0 13.5 6V5Zm.5-.5a.5.5 0 0 1-.5.5v1A1.5 1.5 0 0 0 15 4.5h-1Zm-.5-.5a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 13.5 3v1Zm0-1A1.5 1.5 0 0 0 12 4.5h1a.5.5 0 0 1 .5-.5V3Zm0 8a.5.5 0 0 1-.5-.5h-1a1.5 1.5 0 0 0 1.5 1.5v-1Zm.5-.5a.5.5 0 0 1-.5.5v1a1.5 1.5 0 0 0 1.5-1.5h-1Zm-.5-.5a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 13.5 9v1Zm0-1a1.5 1.5 0 0 0-1.5 1.5h1a.5.5 0 0 1 .5-.5V9Zm-6 5a.5.5 0 0 1-.5-.5H6A1.5 1.5 0 0 0 7.5 15v-1Zm.5-.5a.5.5 0 0 1-.5.5v1A1.5 1.5 0 0 0 9 13.5H8Zm-.5-.5a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 7.5 12v1Zm0-1A1.5 1.5 0 0 0 6 13.5h1a.5.5 0 0 1 .5-.5v-1Zm-6-1a.5.5 0 0 1-.5-.5H0A1.5 1.5 0 0 0 1.5 12v-1Zm.5-.5a.5.5 0 0 1-.5.5v1A1.5 1.5 0 0 0 3 10.5H2Zm-.5-.5a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 1.5 9v1Zm0-1A1.5 1.5 0 0 0 0 10.5h1a.5.5 0 0 1 .5-.5V9Zm0-4a.5.5 0 0 1-.5-.5H0A1.5 1.5 0 0 0 1.5 6V5Zm.5-.5a.5.5 0 0 1-.5.5v1A1.5 1.5 0 0 0 3 4.5H2ZM1.5 4a.5.5 0 0 1 .5.5h1A1.5 1.5 0 0 0 1.5 3v1Zm0-1A1.5 1.5 0 0 0 0 4.5h1a.5.5 0 0 1 .5-.5V3Z"
+        />
+      </svg>
+    ),
+  },
+  Helm: {
+    title: "Helm",
+    ariaLabel: "Helm icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 16 16"
+      >
+        <path
+          fill="currentColor"
+          d="m4.807 4.377.016.015-.147.147a11.93 11.93 0 0 0-.314.324c-.197.213-.372.452-.389.52h.91a.132.132 0 0 0 .109-.049A4.1 4.1 0 0 1 7.778 4a4.14 4.14 0 0 1 3.317 1.336.112.112 0 0 0 .089.047l.595-.001h.295a.31.31 0 0 0 .047-.009 4.982 4.982 0 0 0-.745-.864l.013-.013a3.006 3.006 0 0 0 .72-.823c.135-.213.237-.446.3-.69a.838.838 0 0 0 .022-.29c-.013-.132-.086-.243-.262-.222a.484.484 0 0 0-.103.023 1.058 1.058 0 0 0-.33.193 3.043 3.043 0 0 0-.922 1.31l-.014.03-.006.011a6.106 6.106 0 0 1-.16-.093c-.122-.07-.24-.14-.364-.202-.18-.09-.364-.17-.553-.24a5.217 5.217 0 0 0-.982-.247c-.061-.01-.122-.02-.183-.032a3.15 3.15 0 0 0 .038-1.482 1.486 1.486 0 0 0-.147-.41.484.484 0 0 0-.163-.19.197.197 0 0 0-.225.002.44.44 0 0 0-.089.08.877.877 0 0 0-.153.284c-.068.197-.11.402-.125.61a3.15 3.15 0 0 0 .083 1.073l.004.023a4.938 4.938 0 0 0-2.268.696l-.009-.007a.029.029 0 0 1-.007-.007 3.15 3.15 0 0 0-.62-.96 1.752 1.752 0 0 0-.494-.397.765.765 0 0 0-.16-.062c-.177-.043-.298.049-.31.23l-.001.023v.016c-.007.14.028.272.074.402.166.464.439.856.792 1.196l.035.033zm6.569 7.125c.249-.23.474-.485.67-.76l-.004-.017h-.294c-.2 0-.4.001-.6-.002a.216.216 0 0 0-.168.069c-.95.928-2.086 1.328-3.405 1.2a3.918 3.918 0 0 1-1.392-.414 4.113 4.113 0 0 1-1.103-.809.13.13 0 0 0-.104-.045l-.616.001h-.267a.383.383 0 0 0-.064.005c.23.323.497.617.797.877a.105.105 0 0 1-.01.018l-.025.023-.014.014a3.15 3.15 0 0 0-.742 1.057 1.552 1.552 0 0 0-.121.4.555.555 0 0 0 .02.294.199.199 0 0 0 .15.125c.062.01.125.004.183-.02a1.07 1.07 0 0 0 .337-.194c.376-.313.673-.71.865-1.16l.02-.044.011-.024a4.93 4.93 0 0 0 2.267.7l.003.01c.002.005.003.008.002.01a2.645 2.645 0 0 0-.091.582c-.016.248-.003.497.038.741.032.18.076.356.162.517.03.057.067.11.11.158.108.118.241.117.353.002a.375.375 0 0 0 .065-.084c.044-.079.083-.16.117-.244.07-.203.112-.415.124-.63a3.057 3.057 0 0 0-.092-1.059l-.004-.022c.268-.03.533-.08.793-.15.255-.072.505-.163.747-.273a5.44 5.44 0 0 0 .7-.392l.009.022.011.03c.184.512.503.964.923 1.308.08.064.165.118.257.163.06.029.123.048.19.056a.202.202 0 0 0 .228-.14.418.418 0 0 0 .02-.104.969.969 0 0 0-.061-.403 3.032 3.032 0 0 0-.943-1.348l-.052-.044zm-.556-4.965.072.001c.053.001.105.003.156-.002.084-.008.144.022.205.079.266.245.534.488.802.73l.264.24c.011.01.024.02.037.033l.022.019.02-.018.04-.035 1.1-1.002a.152.152 0 0 1 .117-.046 3.946 3.946 0 0 0 .276.001v2.95a6.801 6.801 0 0 1-.774.002V7.995l-.015-.007-.763.696-.769-.693-.015.005v1.499h-.768c-.014-.05-.02-2.79-.006-2.958zm-6.323.003h-.763v1.046a9.595 9.595 0 0 1-.95-.004v-1.04h-.77v2.947c.064.015.716.014.774-.004V8.373h.946v.743c0 .126-.001.253.002.38H4.5c.016-.072.013-2.906-.003-2.955zm.915 2.954V6.55c.045-.014 1.723-.02 1.833-.005v.627l-.015.002h-.008a.578.578 0 0 1-.052.003h-.98v.489h.934v.65h-.926a3.5 3.5 0 0 0-.007.55l.023.001a.44.44 0 0 0 .052.003h.979v.625H5.412zM8.123 6.54c-.014.055-.017 2.876-.003 2.955h1.832v-.772h-.985l-.053-.003-.03-.002V6.541h-.761z"
+        />
+      </svg>
+    ),
+  },
+  HTML: {
+    title: "HTML",
+    ariaLabel: "HTML icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="M.5.5V0a.5.5 0 0 0-.498.542L.5.5Zm14 0 .498.042A.5.5 0 0 0 14.5 0v.5Zm-1 12 .158.474a.5.5 0 0 0 .34-.433L13.5 12.5Zm-6 2-.158.474a.499.499 0 0 0 .316 0L7.5 14.5Zm-6-2-.498.041a.5.5 0 0 0 .34.433L1.5 12.5Zm3-9V3H4v.5h.5Zm0 3H4V7h.5v-.5Zm6 0h.5V6h-.5v.5Zm0 3 .158.474L11 9.86V9.5h-.5Zm-3 1-.158.474.158.053.158-.053L7.5 10.5Zm-3-1H4v.36l.342.114L4.5 9.5ZM.5 1h14V0H.5v1ZM14.002.458l-1 12 .996.083 1-12-.996-.083Zm-.66 11.568-6 2 .316.948 6-2-.316-.948Zm-5.684 2-6-2-.316.948 6 2 .316-.948Zm-5.66-1.567-1-12-.996.083 1 12 .996-.083ZM11 3H4.5v1H11V3Zm-7 .5v3h1v-3H4ZM4.5 7h6V6h-6v1Zm5.5-.5v3h1v-3h-1Zm.342 2.526-3 1 .316.948 3-1-.316-.948Zm-2.684 1-3-1-.316.948 3 1 .316-.948ZM5 9.5V8H4v1.5h1Z"
+        />
+      </svg>
+    ),
+  },
+  JavaScript: {
+    title: "JavaScript",
+    ariaLabel: "JavaScript icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M12.5 8v-.167c0-.736-.597-1.333-1.333-1.333H10a1.5 1.5 0 1 0 0 3h1a1.5 1.5 0 0 1 0 3h-1A1.5 1.5 0 0 1 8.5 11m-2-5v5a1.5 1.5 0 0 1-3 0M.5.5h14v14H.5V.5Z"
+        />
+      </svg>
+    ),
+  },
+  Jest: {
+    title: "Jest",
+    ariaLabel: "Jest icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 32 32"
+      >
+        <path d="M29.667 15.76a4.155 4.155 0 0 0-3.104-4.015L30.547 0h-19.74l3.996 11.787a4.156 4.156 0 0 0-2.996 3.979c0 1.391.693 2.625 1.751 3.385a10.813 10.813 0 0 1-1.443 1.656 10.92 10.92 0 0 1-3.401 2.199c-1.115-.751-1.593-2.079-1.161-3.26 4.531-1.412 3.531-8.089-1.208-8.109a4.154 4.154 0 0 0-4.151 4.151c0 1.131.457 2.167 1.203 2.911-.068.125-.131.261-.199.396-.619 1.281-1.323 2.724-1.593 4.344-.536 3.245.344 5.849 2.469 7.323a6.778 6.778 0 0 0 3.927 1.24c2.432 0 4.907-1.224 7.297-2.412 1.708-.839 3.468-1.719 5.197-2.145.641-.156 1.308-.249 2.011-.355 1.427-.203 2.901-.416 4.224-1.187a5.966 5.966 0 0 0 2.911-4.12 6.144 6.144 0 0 0-.593-3.828c.401-.641.62-1.385.62-2.193zm-1.807 0a2.336 2.336 0 0 1-2.333 2.333c-2.209-.005-3.177-2.791-1.444-4.167l.005-.009c.079-.057.163-.115.24-.168 0 0 .016 0 .016-.009.036-.021.072-.047.109-.068.011 0 .016-.005.025-.005.037-.021.084-.036.131-.057s.093-.036.131-.057c.009 0 .015-.005.025-.005.037-.011.079-.031.115-.036.005 0 .027-.011.037-.011.047-.011.083-.021.129-.027h.005l.141-.031c.009 0 .025 0 .036-.011.036 0 .073-.011.115-.011h.041c.047 0 .093-.005.151-.005h.12c.037 0 .068 0 .104.005h.016c.073.011.151.021.224.043a2.347 2.347 0 0 1 1.86 2.296zM13.391 1.855h14.573l-3.344 9.864c-.141.027-.276.073-.417.12l-3.52-7.177-3.532 7.131a3.367 3.367 0 0 0-.437-.099zm6.625 13.098a4.117 4.117 0 0 0-1.251-2.224l1.917-3.869 1.937 3.952a4.151 4.151 0 0 0-1.161 2.141zm-4.641-1.464a.845.845 0 0 1 .131-.025h.025c.037-.011.073-.011.109-.021h.037c.036 0 .068-.011.104-.011h.359c.037 0 .068.011.095.011.02 0 .025 0 .047.011.036.011.063.011.099.016.011 0 .021 0 .041.009l.125.027h.011c.036.011.079.021.115.041.005 0 .016.005.036.005.027.011.063.021.095.036.004 0 .015.011.025.011.036.021.073.032.109.047h.011a.425.425 0 0 1 .12.068h.011c.036.016.072.041.109.063.009 0 .009.011.02.011.037.016.063.047.099.063l.011.011c.109.083.213.176.319.271l.004.005c.417.437.647 1.015.641 1.613-.099 3.011-4.568 3.011-4.667 0-.005-1.067.724-2 1.76-2.26zm-9.12-.068a2.336 2.336 0 0 1 0 4.672c-1.285 0-2.333-1.047-2.333-2.333s1.048-2.339 2.333-2.339zm21.552 8.038a4.114 4.114 0 0 1-2.009 2.844c-1.011.583-2.256.771-3.557.952a22.36 22.36 0 0 0-2.188.391c-1.927.475-3.781 1.396-5.579 2.287-2.296 1.141-4.463 2.213-6.473 2.213a4.873 4.873 0 0 1-2.875-.916c-2.037-1.407-1.937-4.047-1.693-5.495.219-1.355.839-2.62 1.432-3.833.043-.073.068-.141.105-.213.203.072.416.129.64.167-.355 1.963.645 3.995 2.593 4.995l.349.181.38-.135c1.62-.579 3.125-1.511 4.448-2.76a12.28 12.28 0 0 0 1.927-2.292 4.157 4.157 0 0 0 4.652-3.057h1.557a4.143 4.143 0 0 0 4.011 3.104c.771 0 1.484-.213 2.093-.573.24.693.317 1.417.187 2.141z" />
+      </svg>
+    ),
+  },
+  JIRA: {
+    title: "JIRA",
+    ariaLabel: "JIRA icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 32 32"
+      >
+        <title>{"jira"}</title>
+        <path d="M29.762 1.004H15.319a6.517 6.517 0 0 0 6.517 6.517h2.66v2.571a6.517 6.517 0 0 0 6.498 6.512h.001V2.256a1.25 1.25 0 0 0-1.232-1.251h-.001zm-7.146 7.194H8.173a6.517 6.517 0 0 0 6.517 6.516h2.66v2.572a6.522 6.522 0 0 0 6.517 6.516V9.45a1.25 1.25 0 0 0-1.251-1.251zm-7.152 7.193H1.004a6.52 6.52 0 0 0 6.521 6.517h2.679v2.57a6.517 6.517 0 0 0 6.514 6.517V16.647c0-.694-.562-1.256-1.256-1.256z" />
+      </svg>
+    ),
+  },
+  JQuery: {
+    title: "jQuery",
+    ariaLabel: "JQuery icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlSpace="preserve"
+        width={800}
+        height={800}
+        viewBox="0 0 512 512"
+      >
+        <path d="M499.161 364.63s-63.554 171.386-282.354 131.488c-3.443-.641-6.745-1.838-10.089-2.849-.782-.229-1.572-.479-2.354-.703a302.903 302.903 0 0 1-9.769-3.197c-.37-.129-.723-.25-1.098-.383a326.662 326.662 0 0 1-8.762-3.244c-.757-.295-1.51-.582-2.263-.89a273.273 273.273 0 0 1-9.277-3.859c-.565-.249-1.11-.499-1.684-.752a317.762 317.762 0 0 1-7.959-3.685c-.682-.346-1.368-.666-2.055-.99-.374-.187-.777-.374-1.16-.565-1.696-.848-3.352-1.729-5.023-2.604-1.289-.665-2.599-1.313-3.88-2a329.103 329.103 0 0 1-7.02-3.938c-.861-.491-1.73-.965-2.583-1.46a332.068 332.068 0 0 1-9.344-5.727c-.869-.557-1.729-1.147-2.586-1.722-2.004-1.297-4.009-2.595-5.972-3.95-.208-.142-.407-.291-.611-.437-.761-.516-1.484-1.044-2.241-1.593a266.52 266.52 0 0 1-6.15-4.408 512.55 512.55 0 0 1-2.932-2.229 321.13 321.13 0 0 1-5.481-4.284c-.902-.723-1.817-1.438-2.719-2.17a341.36 341.36 0 0 1-6.575-5.548c-.408-.357-.84-.707-1.256-1.081-.042-.017-.071-.05-.117-.088a271.49 271.49 0 0 1-7.236-6.591c-.719-.657-1.435-1.36-2.158-2.05a314.61 314.61 0 0 1-5.277-5.178c-.72-.707-1.431-1.422-2.121-2.138a323.508 323.508 0 0 1-6.616-6.97c-.079-.107-.158-.17-.253-.274-.038-.041-.092-.096-.129-.142a305.821 305.821 0 0 1-6.508-7.427 228.322 228.322 0 0 1-1.78-2.133 331.47 331.47 0 0 1-4.674-5.743 360.112 360.112 0 0 1-1.9-2.42 257.465 257.465 0 0 1-5.057-6.721c-.308-.424-.636-.84-.932-1.256a1.41 1.41 0 0 1-.158-.216 283.24 283.24 0 0 1-5.547-8.001c-.561-.824-1.089-1.672-1.63-2.521a240.738 240.738 0 0 1-3.826-5.988c-.648-1.031-1.264-2.066-1.888-3.103a243.51 243.51 0 0 1-3.298-5.655c-.611-1.073-1.214-2.121-1.805-3.223-.329-.582-.686-1.182-1.006-1.792-.562-1.016-1.069-2.075-1.618-3.111-.848-1.626-1.709-3.26-2.528-4.915C-6.655 275.1-12.294 190.5 32.95 125.471l35.314-44.789c-39.843 77.007-21.936 170.778 27.908 238.76 1.8 2.475 3.643 4.916 5.526 7.356.583.736 1.168 1.477 1.751 2.184a284.177 284.177 0 0 0 4.786 5.918c.583.686 1.152 1.359 1.747 2.066a252.343 252.343 0 0 0 6.646 7.564c.108.12.212.229.316.345a303.605 303.605 0 0 0 6.558 6.933c.69.698 1.393 1.389 2.096 2.088a206.2 206.2 0 0 0 5.344 5.235c.686.682 1.376 1.339 2.088 2.017a358.359 358.359 0 0 0 7.435 6.758c.262.229.532.44.773.652a336.406 336.406 0 0 0 7.074 5.96c.832.678 1.68 1.355 2.528 2.029a287.085 287.085 0 0 0 5.681 4.437c.911.699 1.846 1.41 2.791 2.108a342.403 342.403 0 0 0 6.288 4.55c.719.486 1.418 1.015 2.146 1.505a237.617 237.617 0 0 0 6.654 4.425c.84.558 1.672 1.14 2.52 1.681a337.537 337.537 0 0 0 9.357 5.73c.811.466 1.622.902 2.42 1.352a246.141 246.141 0 0 0 7.19 4.029c1.314.731 2.666 1.393 3.996 2.096 1.988 1.023 3.976 2.063 5.988 3.044.761.374 1.501.724 2.246 1.077a371.386 371.386 0 0 0 7.743 3.605c.616.267 1.214.545 1.834.811 3.032 1.327 6.096 2.6 9.186 3.817.794.316 1.589.62 2.387.937a295.645 295.645 0 0 0 8.508 3.156c.445.158.89.32 1.322.478a291.045 291.045 0 0 0 9.648 3.136c.827.258 1.655.508 2.503.757 3.318.981 6.616 2.042 10.018 2.812 89.773 20.431 184.517 2.083 237.895-67.46zM182.44 254.758c4.907 7.024 10.338 15.386 16.846 21.031 2.354 2.598 4.824 5.13 7.34 7.621.632.648 1.285 1.272 1.938 1.908a215.525 215.525 0 0 0 7.494 6.97c.108.079.204.188.308.263.029.033.054.041.075.074 2.845 2.487 5.781 4.887 8.758 7.244.662.508 1.318 1.036 1.988 1.552 2.99 2.291 6.039 4.545 9.166 6.69.096.066.17.134.266.2 1.376.956 2.782 1.846 4.175 2.761.674.433 1.31.898 1.983 1.306a185.394 185.394 0 0 0 6.795 4.167c.325.2.641.366.961.553a250.168 250.168 0 0 0 6.005 3.356c.698.391 1.414.74 2.129 1.114 1.389.724 2.786 1.468 4.2 2.171.212.1.433.199.637.287a199.91 199.91 0 0 0 8.762 4.079c.632.283 1.289.532 1.934.799a224.016 224.016 0 0 0 7.178 2.915c1.022.387 2.059.757 3.073 1.127 2.191.815 4.383 1.556 6.595 2.296 1.003.319 1.984.648 2.986.956 3.144.973 6.254 2.212 9.535 2.753C472.501 366.942 511.5 236.873 511.5 236.873c-35.16 50.633-103.243 74.791-175.897 55.936a209.508 209.508 0 0 1-9.56-2.77 96.802 96.802 0 0 1-2.883-.928 197.877 197.877 0 0 1-6.694-2.308c-1.011-.37-2.025-.74-3.02-1.123a228.945 228.945 0 0 1-7.198-2.902c-.653-.283-1.306-.533-1.942-.824a201.833 201.833 0 0 1-8.812-4.094c-1.493-.729-2.953-1.499-4.425-2.254l-2.549-1.312a213.673 213.673 0 0 1-5.594-3.154c-.437-.252-.894-.491-1.339-.755a228.56 228.56 0 0 1-6.77-4.134c-.687-.438-1.352-.915-2.038-1.366a239.846 239.846 0 0 1-4.392-2.902c-3.11-2.146-6.138-4.399-9.128-6.678-.69-.543-1.376-1.073-2.059-1.616-31.796-25.111-57-59.423-68.994-98.33-12.555-40.352-9.847-85.647 11.914-122.406l-26.735 37.738c-32.715 47.085-30.939 110.113-5.419 159.91 4.286 8.349 9.097 16.445 14.474 24.157zm178.06-58.252c1.393.52 2.782.981 4.191 1.458.616.189 1.231.401 1.851.584 1.988.626 3.963 1.352 6.018 1.738 93.284 18.023 118.58-47.869 125.317-57.566-22.165 31.91-59.405 39.57-105.111 28.481-3.605-.881-7.572-2.185-11.058-3.416a135.451 135.451 0 0 1-13.124-5.456 136.046 136.046 0 0 1-23.025-14.045c-40.891-31.029-66.291-90.221-39.61-138.425L291.51 29.76c-19.287 28.386-21.179 63.638-7.801 94.991 14.114 33.274 43.036 59.381 76.791 71.755z" />
+      </svg>
+    ),
+  },
+  Kubernetes: {
+    title: "Kubernetes",
+    ariaLabel: "Kubernetes icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 16 16"
+      >
+        <path
+          fill="currentColor"
+          fillRule="evenodd"
+          d="M5.374 15c-.335 0-.66-.153-.874-.431l-3.254-4.172a1.15 1.15 0 0 1-.214-.978l1.165-5.207c.074-.345.298-.623.605-.776l4.715-2.32A1.05 1.05 0 0 1 8.002 1c.168 0 .335.038.485.115L13.2 3.426c.308.153.532.432.606.777l1.165 5.207c.074.345 0 .7-.214.978L11.5 14.566c-.214.268-.54.434-.875.434h-5.25zm7.718-5.835.031.008a.308.308 0 0 1 .26.371.306.306 0 0 1-.396.223h-.004l-.003-.001-.003-.001-.03-.007-.05-.01a2.548 2.548 0 0 1-.274-.106 2.87 2.87 0 0 0-.533-.156.242.242 0 0 0-.171.063 4.76 4.76 0 0 0-.131-.023 3.972 3.972 0 0 1-1.764 2.212c.015.042.032.083.051.123a.239.239 0 0 0-.023.18 2.8 2.8 0 0 0 .271.484c.06.078.114.16.164.244l.028.057.012.025a.306.306 0 0 1-.381.44.307.307 0 0 1-.172-.18 2.608 2.608 0 0 0-.01-.02l-.028-.058a2.545 2.545 0 0 1-.089-.28 2.835 2.835 0 0 0-.21-.512.242.242 0 0 0-.156-.095 5.926 5.926 0 0 1-.03-.053l-.035-.064a3.97 3.97 0 0 1-2.824-.007l-.069.125a.249.249 0 0 0-.132.064c-.104.17-.184.355-.237.548a2.525 2.525 0 0 1-.088.28l-.025.05-.013.027v.001a.307.307 0 1 1-.553-.261l.014-.03.026-.052c.05-.085.104-.166.164-.244.108-.156.2-.322.277-.496a.302.302 0 0 0-.028-.173l.056-.133A3.972 3.972 0 0 1 4.22 9.532l-.134.023a.34.34 0 0 0-.176-.062 2.872 2.872 0 0 0-.533.156c-.09.04-.181.075-.274.105l-.05.011-.03.007H3.02l-.002.002h-.005a.308.308 0 0 1-.397-.349.306.306 0 0 1 .261-.245l.004-.001h.003l.006-.002c.024-.006.054-.014.076-.018.097-.013.195-.021.293-.023a2.98 2.98 0 0 0 .549-.09.422.422 0 0 0 .131-.133l.128-.037a3.938 3.938 0 0 1 .624-2.752l-.097-.087a.338.338 0 0 0-.062-.176 2.854 2.854 0 0 0-.455-.319 2.557 2.557 0 0 1-.254-.148 1.129 1.129 0 0 1-.063-.05l-.004-.004a.323.323 0 0 1-.076-.45.295.295 0 0 1 .244-.107.365.365 0 0 1 .213.08l.022.017c.016.013.034.026.046.037.071.067.139.139.202.213.125.137.263.262.412.372a.24.24 0 0 0 .182.018l.11.078a3.938 3.938 0 0 1 2.552-1.224l.007-.129a.332.332 0 0 0 .1-.157 2.844 2.844 0 0 0-.034-.554 2.555 2.555 0 0 1-.042-.29V3.158A.306.306 0 0 1 8 2.82a.308.308 0 0 1 .306.337v.087a2.53 2.53 0 0 1-.041.29 2.85 2.85 0 0 0-.035.553.242.242 0 0 0 .1.153v.007l.007.13c.967.087 1.87.522 2.54 1.223l.116-.083a.34.34 0 0 0 .186-.02 2.8 2.8 0 0 0 .412-.373 2.53 2.53 0 0 1 .202-.213l.051-.04.017-.014a.308.308 0 0 1 .472.388.307.307 0 0 1-.09.09.146.146 0 0 0-.025.02l-.043.033a2.549 2.549 0 0 1-.254.148 2.865 2.865 0 0 0-.455.32.24.24 0 0 0-.058.172 4.458 4.458 0 0 1-.05.044l-.058.053c.542.806.769 1.783.637 2.745l.123.036a.342.342 0 0 0 .132.132c.18.048.364.078.55.09.097.003.195.01.292.024l.058.013zm-2.875-3.1-1.308.925-.004-.002a.27.27 0 0 1-.43-.205v-.001l-.091-1.598a3.183 3.183 0 0 1 1.833.882zM7.754 7.818h.492l.306.381-.11.476L8 8.886l-.443-.213-.11-.475.307-.381zM7.29 5.24c.107-.024.216-.043.326-.056l-.09 1.6-.008.004a.268.268 0 0 1-.293.256.27.27 0 0 1-.135-.05l-.002.001-1.316-.93c.419-.41.945-.696 1.518-.825zM5.296 6.663l1.201 1.071-.001.007a.269.269 0 0 1-.106.462l-.001.005-1.54.443a3.134 3.134 0 0 1 .447-1.988zm1.608 2.846-.612 1.474a3.16 3.16 0 0 1-1.27-1.586L6.6 9.13l.003.003a.265.265 0 0 1 .18.029.27.27 0 0 1 .117.341l.004.006zm1.806 1.896c-.572.13-1.17.1-1.726-.088l.777-1.4h.001a.27.27 0 0 1 .475-.001h.006l.779 1.402a3.286 3.286 0 0 1-.312.087zm1.004-.416L9.096 9.5l.001-.003a.269.269 0 0 1 .296-.37l.003-.004 1.593.269a3.147 3.147 0 0 1-1.275 1.597zm1.442-2.343L9.61 8.201l-.002-.006a.27.27 0 0 1-.185-.343.27.27 0 0 1 .08-.12L9.5 7.73l1.195-1.067c.366.594.527 1.29.46 1.983z"
+          clipRule="evenodd"
+        />
+      </svg>
+    ),
+  },
+  MaterialUI: {
+    title: "MaterialUI",
+    ariaLabel: "MaterialUI icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <title>{"material_ui"}</title>
+        <path fill="none" d="M0 0h24v24H0z" />
+        <path d="M8 16.61v-1.24l6-3.46V7.23l-5 2.89-5-2.89V13l-1 .58L2 13V5l1.07-.62L9 7.81l3.93-2.27 2-1.16L16 5v8.06L10.92 16 15 18.33l5-2.9V11l1-.58 1 .58v5.58l-7 4.06-7-4m14-6.89-1 .58-1-.58V8.58L21 8l1 .58Z" />
+      </svg>
+    ),
+  },
+  NextJS: {
+    title: "NextJS",
+    ariaLabel: "NextJS icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="m4.5 4.5.405-.293A.5.5 0 0 0 4 4.5h.5Zm3 9.5A6.5 6.5 0 0 1 1 7.5H0A7.5 7.5 0 0 0 7.5 15v-1ZM14 7.5A6.5 6.5 0 0 1 7.5 14v1A7.5 7.5 0 0 0 15 7.5h-1ZM7.5 1A6.5 6.5 0 0 1 14 7.5h1A7.5 7.5 0 0 0 7.5 0v1Zm0-1A7.5 7.5 0 0 0 0 7.5h1A6.5 6.5 0 0 1 7.5 1V0ZM5 12V4.5H4V12h1Zm-.905-7.207 6.5 9 .81-.586-6.5-9-.81.586ZM10 4v6h1V4h-1Z"
+        />
+      </svg>
+    ),
+  },
+  NodeJS: {
+    title: "NodeJS",
+    ariaLabel: "NodeJS icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 16 16"
+      >
+        <path
+          fill="currentColor"
+          d="M7.586.102a.75.75 0 0 1 .756 0l6.214 3.625a.75.75 0 0 1 .373.648v7.25a.75.75 0 0 1-.373.648l-6.214 3.625a.75.75 0 0 1-.756 0l-1.553-.906a.75.75 0 0 1 .756-1.296l1.175.686 5.465-3.188V4.806L7.964 1.618 2.5 4.806v6.388l1.117.652c.344.11.77.137 1.069.005a.75.75 0 0 0 .364-.34c.104-.188.2-.501.2-1.011V5.25a.75.75 0 0 1 1.5 0v5.25c0 .69-.129 1.272-.39 1.74a2.25 2.25 0 0 1-1.068.983c-.806.356-1.697.21-2.242.014a.75.75 0 0 1-.124-.058l-1.554-.906A.75.75 0 0 1 1 11.625v-7.25a.75.75 0 0 1 .372-.648L7.586.102Zm.66 5.181c.397-.472 1.017-.783 1.858-.783.78 0 1.375.268 1.778.693a.75.75 0 1 1-1.09 1.03C10.703 6.132 10.52 6 10.103 6c-.442 0-.624.148-.709.248a.642.642 0 0 0-.145.377c0 .083.036.247.145.377.085.1.267.248.71.248a.75.75 0 0 1 .212.03c.62.09 1.092.368 1.416.753.372.443.496.967.496 1.342 0 .375-.124.899-.496 1.342-.397.472-1.017.783-1.858.783-.78 0-1.375-.268-1.777-.693a.75.75 0 0 1 1.09-1.03c.087.092.271.223.687.223.442 0 .625-.148.71-.248a.642.642 0 0 0 .144-.377.642.642 0 0 0-.145-.377c-.084-.1-.267-.248-.709-.248a.75.75 0 0 1-.213-.03c-.62-.09-1.092-.368-1.415-.753a2.137 2.137 0 0 1-.497-1.342c0-.375.124-.899.497-1.342Z"
+        />
+      </svg>
+    ),
+  },
+  PHP: {
+    title: "PHP",
+    ariaLabel: "PHP icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlSpace="preserve"
+        width={800}
+        height={800}
+        viewBox="0 0 512 512"
+      >
+        <path d="M171.844 204.374c-11.137-12.748-28.856-19.123-53.146-19.123H37.96L.5 377.99h41.984l9.96-51.241h35.963c15.869 0 28.923-1.663 39.173-5.003 10.247-3.33 19.562-8.92 27.945-16.767 7.037-6.467 12.725-13.599 17.087-21.4 4.354-7.797 7.448-16.401 9.278-25.812 4.443-22.848 1.09-40.643-10.046-53.393zm-33.351 50.449c-2.903 14.917-8.492 25.563-16.775 31.941-8.288 6.38-20.897 9.569-37.822 9.569H58.354l15.678-80.667H102.8c15.952 0 26.582 2.943 31.896 8.832 5.31 5.892 6.579 15.999 3.797 30.325zm199.335-17.764-17.429 89.69h-42.317l16.572-85.278c1.884-9.702 1.193-16.32-2.084-19.847-3.272-3.529-10.242-5.296-20.9-5.296h-33.289l-21.458 110.421h-41.656l37.46-192.739h41.656l-9.959 51.241h37.111c23.346 0 39.452 4.077 48.317 12.218 8.866 8.146 11.519 21.344 7.976 39.59zm161.726-32.685c-11.137-12.748-28.856-19.123-53.142-19.123h-80.738l-37.46 192.739h41.984l9.96-51.241h35.963c15.869 0 28.918-1.663 39.169-5.003 10.247-3.33 19.562-8.92 27.945-16.767 7.036-6.467 12.729-13.599 17.088-21.4 4.354-7.797 7.447-16.401 9.277-25.812 4.442-22.848 1.094-40.643-10.046-53.393zm-33.348 50.449c-2.902 14.917-8.491 25.563-16.779 31.941-8.284 6.38-20.896 9.569-37.822 9.569h-25.537l15.678-80.667h28.765c15.952 0 26.581 2.943 31.899 8.832 5.31 5.892 6.574 15.999 3.796 30.325z" />
+      </svg>
+    ),
+  },
+  Python: {
+    title: "Python",
+    ariaLabel: "Python icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M6 2.5h1M4.5 4V1.5a1 1 0 0 1 1-1h4a1 1 0 0 1 1 1v5a1 1 0 0 1-1 1h-4a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1h4a1 1 0 0 0 1-1V11M8 4.5H1.5a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1h3m2.5-1h6.5a1 1 0 0 0 1-1v-5a1 1 0 0 0-1-1h-3m-2.5 9h1"
+        />
+      </svg>
+    ),
+  },
+  React2: {
+    title: "React",
+    ariaLabel: "React icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M14.5 7.584c0 1.657-3.134 3-7 3s-7-1.343-7-3 3.134-3 7-3 7 1.343 7 3Z"
+        />
+        <path
+          stroke="currentColor"
+          d="M4.166 13.739c1.457.79 4.13-1.327 5.972-4.726 1.841-3.4 2.153-6.795.696-7.584-1.457-.79-4.13 1.327-5.972 4.726-1.841 3.4-2.153 6.795-.696 7.584Z"
+        />
+        <path
+          stroke="currentColor"
+          d="M10.834 13.739c-1.457.79-4.13-1.327-5.972-4.726-1.841-3.4-2.153-6.795-.696-7.584 1.457-.79 4.13 1.327 5.972 4.726 1.841 3.4 2.153 6.795.696 7.584Z"
+        />
+        <path
+          stroke="currentColor"
+          d="M6.5 7.584a1 1 0 1 0 2 0 1 1 0 0 0-2 0Z"
+        />
+      </svg>
+    ),
+  },
+  Redux: {
+    title: "Redux",
+    ariaLabel: "Redux icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 24 24"
+      >
+        <path
+          fill="currentColor"
+          d="M16.63 16.563c.885-.092 1.557-.863 1.527-1.788a1.723 1.723 0 0 0-1.71-1.665h-.062c-.947.03-1.68.832-1.65 1.788.032.463.215.863.49 1.14-1.039 2.067-2.627 3.577-5.01 4.841-1.618.864-3.298 1.172-4.977.956-1.375-.185-2.444-.802-3.116-1.819-.977-1.51-1.068-3.145-.244-4.779.58-1.171 1.497-2.035 2.077-2.466-.122-.401-.305-1.08-.397-1.573C-.871 14.436-.412 18.814.93 20.88c1.008 1.542 3.054 2.498 5.315 2.498.61 0 1.221-.062 1.832-.216 3.91-.77 6.872-3.114 8.552-6.598Zm5.375-3.823c-2.321-2.744-5.742-4.255-9.651-4.255h-.489a1.677 1.677 0 0 0-1.496-.925h-.062c-.946.031-1.68.833-1.649 1.789.03.925.794 1.664 1.71 1.664h.062a1.721 1.721 0 0 0 1.496-1.048h.55c2.321 0 4.52.678 6.505 2.004 1.527 1.018 2.627 2.343 3.237 3.947.52 1.294.49 2.558-.06 3.638-.856 1.634-2.291 2.528-4.185 2.528-1.221 0-2.382-.37-2.993-.648a18.07 18.07 0 0 1-1.374 1.11c1.313.617 2.657.956 3.94.956 2.932 0 5.1-1.634 5.925-3.268.885-1.788.824-4.871-1.466-7.492ZM6.49 17.087c.03.925.794 1.665 1.71 1.665h.061c.947-.03 1.68-.832 1.65-1.788a1.723 1.723 0 0 0-1.71-1.665h-.062c-.061 0-.153 0-.214.03-1.252-2.096-1.771-4.377-1.588-6.844.122-1.85.733-3.453 1.802-4.779.886-1.14 2.596-1.695 3.757-1.726 3.237-.062 4.611 4.008 4.703 5.642.397.092 1.069.308 1.527.462C17.759 3.09 14.706.5 11.773.5 9.025.5 6.49 2.504 5.482 5.464c-1.405 3.946-.489 7.738 1.222 10.729-.153.216-.245.555-.214.894Z"
+        />
+      </svg>
+    ),
+  },
+  Storybook: {
+    title: "Storybook",
+    ariaLabel: "Storybook icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <title>{"Storybook icon"}</title>
+        <path d="m16.34.24-.12 2.71a.18.18 0 0 0 .29.15l1.06-.8.9.7a.18.18 0 0 0 .28-.14L18.65.1l1.33-.1a1.2 1.2 0 0 1 1.28 1.2v21.6A1.2 1.2 0 0 1 20 24l-16.1-.72a1.2 1.2 0 0 1-1.15-1.16L2 2.32a1.2 1.2 0 0 1 1.13-1.27l13.2-.83.01.02zM13.27 9.3c0 .47 3.16.24 3.59-.08 0-3.2-1.72-4.89-4.86-4.89-3.15 0-4.9 1.72-4.9 4.29 0 4.45 6 4.53 6 6.96 0 .7-.32 1.1-1.05 1.1-.96 0-1.35-.49-1.3-2.16 0-.36-3.65-.48-3.77 0-.27 4.03 2.23 5.2 5.1 5.2 2.79 0 4.97-1.49 4.97-4.18 0-4.77-6.1-4.64-6.1-7 0-.97.72-1.1 1.13-1.1.45 0 1.25.07 1.19 1.87z" />
+      </svg>
+    ),
+  },
+  Trello: {
+    title: "Trello",
+    ariaLabel: "Trello icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 24 24"
+      >
+        <path d="M21.003 0h.04c.814 0 1.55.337 2.075.88l.001.001a2.88 2.88 0 0 1 .88 2.076v.047-.002V21.044c0 .814-.337 1.55-.88 2.075l-.001.001a2.88 2.88 0 0 1-2.076.88h-.043.002H2.955c-.814 0-1.55-.337-2.075-.88l-.001-.001a2.88 2.88 0 0 1-.88-2.076V21v.002V2.96c0-.814.337-1.55.88-2.075L.88.884a2.88 2.88 0 0 1 2.076-.88h.043-.002zM10.448 18.162V4.537c0-.395-.164-.751-.428-1.004a1.386 1.386 0 0 0-.998-.423H4.543c-.395 0-.751.164-1.004.428a1.39 1.39 0 0 0-.429 1.004v13.619l-.001.05c0 .392.165.745.429.994l.001.001c.258.248.609.4.996.4h.023-.001 4.468c.387 0 .738-.153.997-.401.265-.25.43-.603.43-.995l-.001-.052v.003zm10.448-6V4.538c0-.395-.164-.751-.428-1.004a1.386 1.386 0 0 0-1.004-.429h-4.473c-.395 0-.751.164-1.004.428-.265.254-.429.61-.429 1.004v.014-.001 7.608l-.001.05c0 .392.165.745.429.994l.001.001c.258.248.609.4.996.4h.023-.001 4.468c.387 0 .738-.153.997-.401.265-.25.43-.603.43-.995l-.001-.048v.002z" />
+      </svg>
+    ),
+  },
+  TypeScript: {
+    title: "TypeScript",
+    ariaLabel: "TypeScript icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          stroke="currentColor"
+          d="M12.5 8v-.167c0-.736-.597-1.333-1.333-1.333H10a1.5 1.5 0 1 0 0 3h1a1.5 1.5 0 0 1 0 3h-1A1.5 1.5 0 0 1 8.5 11M8 6.5H3m2.5 0V13M.5.5h14v14H.5V.5Z"
+        />
+      </svg>
+    ),
+  },
+  Unity: {
+    title: "Unity",
+    ariaLabel: "Unity icon",
+    icon: () => (
+      <svg
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        viewBox="0 0 32 32"
+      >
+        <title>{"unity"}</title>
+        <path d="m25.94 25.061-5.382-9.06 5.382-9.064 2.598 9.062-2.599 9.06zm-11.994-.87-6.768-6.717h10.759l5.38 9.061-9.372-2.342zm0-16.382 9.371-2.342-5.379 9.061H7.177zm17.05 5.108L27.714 1.004 15.463 4.197l-1.812 3.112-3.68-.027-8.966 8.719 8.967 8.72 3.678-.029 1.817 3.112 12.246 3.192 3.283-11.908-1.864-3.087z" />
+      </svg>
+    ),
+  },
+  VueJS: {
+    title: "VueJS",
+    ariaLabel: "VueJS icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="m7.5 13.5-.432.252a.5.5 0 0 0 .864 0L7.5 13.5Zm4-12 .434.248A.5.5 0 0 0 11.5 1v.5Zm-4 7-.434.248a.5.5 0 0 0 .868 0L7.5 8.5Zm-4-7V1a.5.5 0 0 0-.434.748L3.5 1.5Zm3 0 .447-.224L6.81 1H6.5v.5Zm1 2-.447.224a.5.5 0 0 0 .894 0L7.5 3.5Zm1-2V1h-.309l-.138.276.447.224Zm-8.432.252 7 12 .864-.504-7-12-.864.504Zm7.864 12 7-12-.864-.504-7 12 .864.504Zm3.134-12.5-4 7 .868.496 4-7-.868-.496Zm-3.132 7-4-7-.868.496 4 7 .868-.496ZM3.5 2h3V1h-3v1Zm2.553-.276 1 2 .894-.448-1-2-.894.448Zm1.894 2 1-2-.894-.448-1 2 .894.448ZM8.5 2h3V1h-3v1Z"
+        />
+      </svg>
+    ),
+  },
+  WordPress: {
+    title: "WordPress",
+    ariaLabel: "WordPress icon",
+    icon: () => (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width={800}
+        height={800}
+        fill="none"
+        viewBox="0 0 15 15"
+      >
+        <path
+          fill="currentColor"
+          d="m5.5 13.5-.479.144a.5.5 0 0 0 .957.003L5.5 13.5Zm4 0-.479.144a.5.5 0 0 0 .943.042L9.5 13.5Zm3.53-8.827.465.186-.464-.186Zm-.54-1.931-.3.4.3-.4Zm-.133-.1.3-.4-.3.4Zm-2 2 .4-.3-.4.3Zm.286-2-.3-.4.3.4ZM7.5 14A6.5 6.5 0 0 1 1 7.5H0A7.5 7.5 0 0 0 7.5 15v-1ZM14 7.5A6.5 6.5 0 0 1 7.5 14v1A7.5 7.5 0 0 0 15 7.5h-1ZM7.5 1A6.5 6.5 0 0 1 14 7.5h1A7.5 7.5 0 0 0 7.5 0v1Zm0-1A7.5 7.5 0 0 0 0 7.5h1A6.5 6.5 0 0 1 7.5 1V0ZM2.021 3.644l3 10 .958-.288-3-10-.958.288Zm3.957 10.003 2-6.5-.956-.294-2 6.5.956.294ZM6.02 3.644l3 10 .958-.288-3-10-.958.288Zm6.768-1.302-.132-.1-.6.8.132.1.6-.8Zm-2.832 2.6.643.858.8-.6-.643-.857-.8.6Zm.386-2.7a1.929 1.929 0 0 0-.386 2.7l.8-.6a.929.929 0 0 1 .186-1.3l-.6-.8Zm2.314 0a1.928 1.928 0 0 0-2.314 0l.6.8a.928.928 0 0 1 1.114 0l.6-.8Zm.838 2.617a2.149 2.149 0 0 0-.706-2.517l-.6.8c.416.312.57.863.377 1.345l.929.372ZM2 4h2V3H2v1Zm3 0h3V3H5v1Zm4.964 9.686 3.531-8.827-.928-.372-3.531 8.827.928.372Z"
+        />
+      </svg>
+    ),
+  },
 };
 
 export type IconName = keyof typeof Icons;
