@@ -1,10 +1,39 @@
-const intro =
-  "I am a passionate frontend web developer backed by over 5 years of professional experience. I specialize in JavaScript libraries and frameworks with a strong proficiency in React and TypeScript.";
+import { IconNames } from "../Icons";
 
-const workExperience = [
+const intro =
+  "I am a passionate web developer backed by over 5 years of professional experience. I specialize in JavaScript libraries and frameworks with a strong proficiency in React and TypeScript.";
+
+type WorkExperience = {
+  company: string;
+  title: string;
+  skills: IconNames[];
+  description: string;
+  location: string;
+  startDate: string;
+  endDate: string;
+  highlights: string[];
+};
+
+const workExperience: WorkExperience[] = [
   {
     company: "Auvik Networks",
     title: "Software Developer",
+    skills: [
+      "React",
+      "TypeScript",
+      "HTML",
+      "CSS",
+      "Jest",
+      "Storybook",
+      "GraphQL",
+      "GoLang",
+      "AWS",
+      "Docker",
+      "Kubernetes",
+      "Helm",
+      "JIRA",
+      "Gitlab",
+    ],
     description:
       "Collaborated as a frontend specialist in an agile team with a primary focus on account management and the consolidation of legacy code into a unified platform. Executed a wide range of tasks from minor microservice updates to leading full project lifecycles.",
     location: "Waterloo, Canada (Remote)",
@@ -23,6 +52,23 @@ const workExperience = [
   {
     company: "Bitovi",
     title: "JavaScript Consultant",
+    skills: [
+      "React",
+      "TypeScript",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Jest",
+      "Storybook",
+      "GraphQL",
+      "NextJS",
+      "Redux",
+      "NodeJS",
+      "MaterialUI",
+      "ChakraUI",
+      "JIRA",
+      "GitHub",
+    ],
     description:
       "Engaged as a versatile developer and consultant within Bitovi's React Department, serving an array of diverse clients. Proficiently managed a broad spectrum of frontend React projects from short-term individual assignments to substantial, collaborative endeavors within agile teams.",
     location: "Chicago, USA (Remote)",
@@ -39,6 +85,23 @@ const workExperience = [
   {
     company: "Richmond Day",
     title: "Full Stack Developer",
+    skills: [
+      "JQuery",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Bootstrap",
+      "ASPnet",
+      "Csharp",
+      //"Entity Framework",
+      //"AJAX",
+      "SQL",
+      "WordPress",
+      "PHP",
+      //"Liquid",
+      "JIRA",
+      "GitHub",
+    ],
     description:
       "Collaborated closely with project managers and designers to plan, develop, and maintain websites for an extensive and distinguished clientele, including industry leaders such as MINI, BMW, and LG.",
     location: "Toronto, Canada (Remote)",
@@ -53,20 +116,52 @@ const workExperience = [
   {
     company: "Self Employed",
     title: "Freelance Web Developer",
+    skills: [
+      "JQuery",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Bootstrap",
+      "WordPress",
+      "PHP",
+      "Figma",
+      "Photoshop",
+      "Illustrator",
+    ],
     description:
       "Cultivated client relationships, gathered project requirements, facilitated contract negotiations, and consistently delivered websites, primarily leveraging WordPress and custom themes.",
     location: "Toronto, Canada (Remote)",
-    startDate: "May 2018",
+    startDate: "November 2019",
     endDate: "April 2020",
     highlights: [
       "Managed deployment and hosting of web domains as well as SEO, contact forms, and eCommerce",
       "Worked closely with non-technical clients to gather requirements and create mockups true to their vision and budget",
-      "Delivered websites from conception to production within timelines as short as three days",
+      "Designed, developed, and deployed luxury watch splash page within tight two day deadline",
+      "Refactored code, improved SEO, and established article publication workflow for psychology article website",
+      "Integrated calendar scheduling, PayPal services, and improved UI layout for virtual classroom website",
     ],
   },
   {
     company: "LBC IT Solutions, Inc",
     title: "Junior Web Developer",
+    skills: [
+      "VueJS",
+      "React",
+      "JQuery",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Bootstrap",
+      "ASPnet",
+      "Csharp",
+      //"Entity Framework",
+      "SQL",
+      "WordPress",
+      "PHP",
+      "Trello",
+      "Bitbucket",
+      "Azure",
+    ],
     description:
       "Worked in a small development team to deliver large-scale web applications as well as WordPress websites for clients.",
     location: "Toronto, Canada (Remote)",
@@ -81,6 +176,16 @@ const workExperience = [
   {
     company: "PWC Technology Services, Inc",
     title: "Software Developer Intern",
+    skills: [
+      "JQuery",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Bootstrap",
+      //"AJAX",
+      //"MySQL",
+      "PHP",
+    ],
     description:
       "Collaborated in an agile development team contracted by Scotiabank and worked independently on web applications.",
     location: "Scarborough, Canada",
@@ -121,22 +226,37 @@ const skillHighlights = [
   "Adept with CI/CD deployment workflows",
 ];
 
-const technicalSkills = [
+type TechnicalSkill = {
+  title: string;
+  skills: string[];
+};
+
+const technicalSkills: TechnicalSkill[] = [
   {
     title: "Languages",
-    expert: ["HTML", "CSS", "JavaScript", "TypeScript", "C#", "SQL", "PHP"],
-    advanced: ["GoLang", "Python", "Bash"],
-    intermediate: ["Liquid", "Java", "C++", "C"],
+    skills: [
+      "HTML",
+      "CSS",
+      "JavaScript",
+      "TypeScript",
+      "C#",
+      "SQL",
+      "PHP",
+      "GoLang",
+      "Python",
+      "Bash",
+      "Liquid",
+      "Java",
+      "C++",
+      "C",
+    ],
   },
   {
     title: "Libraries & Frameworks",
-    expert: [
+    skills: [
       "React",
-      "React Hooks",
       "NextJS",
-      "Jest",
-      "Storybook",
-      "Node",
+      "NodeJS",
       "Redux",
       "Scss",
       "Sass",
@@ -147,35 +267,44 @@ const technicalSkills = [
       "jQuery",
       "AJAX",
       "WordPress",
-    ],
-    advanced: [
       "GraphQL",
       "VueJS",
       "RxJS",
       "XState",
-      "Spring",
+      "React-spring",
       "Material UI",
       "Chakra UI",
       "MySQL",
       "MongoDB",
       "Unity",
+      "AngularJS",
+      "BabylonJS",
+      "CanJS",
+      "Colyseus",
     ],
-    intermediate: ["AngularJS, BabylonJS, CanJS"],
   },
   {
-    title: "Tools & Technologies",
-    expert: [
+    title: "Development Tools",
+    skills: [
       "Git",
-      "RESTful API",
+      "RESTful APIs",
       "JIRA",
-      "SEO",
-      "Dev Tools",
-      "JSON",
+      "Webpack",
       "Figma",
+      "SEO",
+      "OpenGraph",
+      "Browser Dev Tools",
+      "JSON",
       "Vim",
     ],
-    advanced: ["Webpack", "AWS", "Docker", "Azure", "Vercel"],
-    intermediate: ["Kubernetes", "Helm", "Postman"],
+  },
+  {
+    title: "Cloud & Deployment",
+    skills: ["Vercel", "AWS", "Docker", "Azure", "Kubernetes", "Helm"],
+  },
+  {
+    title: "Testing & Debugging",
+    skills: ["Jest", "Storybook", "React Testing Library", "Postman"],
   },
 ];
 
@@ -191,7 +320,6 @@ const hobbies = [
 export const ResumeContent = {
   intro,
   workExperience,
-  skillHighlights,
   technicalSkills,
   education,
   hobbies,
