@@ -9,7 +9,7 @@ export const SocialIcons = () => {
         target="_blank"
         rel="noreferrer"
       >
-        <StickyNote size="small" icon="Github" text="github" />
+        <StickyNote size="small" icon="GitHub" text="github" />
       </a>
       <a
         href="https://www.linkedin.com/in/weston-vincze/"
